wip(iter-247): split commands/lint.rs (4005 lines) into lint/ submodules
