+++ crates/hyalo-cli/src/commands/lint.rs
@@ -1,27 +1,16 @@
-/// `hyalo lint` — validate frontmatter properties against the `.hyalo.toml` schema.
-///
-/// Reads each file's frontmatter, applies the type-specific schema (or the
-/// default schema if `type` is absent), and reports violations at two severity
-/// levels:
-///
-///   - **error**  — schema violation (missing required field, wrong value type,
-///     invalid enum value, failed pattern match, empty value on a list-typed
-///     required property)
-///   - **warn**   — soft issue (no `type` property, property not declared in
-///     schema)
-///
-/// Exit code: 0 = clean, 1 = errors found, 2 = internal error.
-use std::borrow::Cow;
-use std::collections::HashSet;
-use std::path::Path;
-
-use anyhow::{Context, Result};
-
-use hyalo_core::frontmatter::{check_mtime, read_mtime, write_frontmatter_within};
-use hyalo_core::scanner;
-use hyalo_core::schema::{PropertyConstraint, SchemaConfig, TypeSchema};
-use hyalo_mdlint::schema::terse_root_cause;
-
+//! `hyalo lint` — validate frontmatter properties against the `.hyalo.toml` schema.
+//!
+//! Reads each file's frontmatter, applies the type-specific schema (or the
+//! default schema if `type` is absent), and reports violations at two severity
+//! levels:
+//!
+//!   - **error**  — schema violation (missing required field, wrong value type,
+//!     invalid enum value, failed pattern match, empty value on a list-typed
+//!     required property)
+//!   - **warn**   — soft issue (no `type` property, property not declared in
+//!     schema)
+//!
+//! Exit code: 0 = clean, 1 = errors found, 2 = internal error.
 // ARCH-2 (iter-226): the schema-lint core (types, per-file validation,
 // auto-fix, constraint helpers) lives in `hyalo_mdlint::schema`. Re-exported
 // here so existing `commands::lint::` call sites keep working.
@@ -35,3971 +24,35 @@ pub use hyalo_mdlint::schema::{
     validate_properties, validate_required_sections,
 };
 
-use crate::output::{CommandOutcome, Format};
-
 // ---------------------------------------------------------------------------
-// Public API
+// Submodules (iteration 247)
 // ---------------------------------------------------------------------------
-
-/// Validate `.hyalo.toml` view definitions and return a pseudo-file lint
-/// result when at least one view looks suspicious.
-///
-/// Current checks:
-/// - Views whose only narrowing mechanism is `fields = ["backlinks"]` or
-///   similar — `fields` controls display columns, not filtering, so such a
-///   view matches every file. The likely intent is `orphan = true`.
-///
-/// Returns `None` when there is nothing to report.
-pub fn validate_views(dir: &Path) -> Option<FileLintResult> {
-    // Keys that actually *narrow* the result set.
-    const NARROWING_KEYS: &[&str] = &[
-        "pattern",
-        "regexp",
-        "properties",
-        "tag",
-        "task",
-        "sections",
-        "file",
-        "glob",
-        "broken_links",
-        "orphan",
-        "dead_end",
-        "title",
-        "language",
-    ];
-
-    let toml_path = dir.join(".hyalo.toml");
-    let contents = std::fs::read_to_string(&toml_path).ok()?;
-    let table: toml::Table = toml::from_str(&contents).ok()?;
-    let Some(toml::Value::Table(views_table)) = table.get("views") else {
-        return None;
-    };
-
-    let mut violations: Vec<Violation> = Vec::new();
-    for (name, value) in views_table {
-        let Some(view_tbl) = value.as_table() else {
-            continue;
-        };
-
-        let has_narrowing = view_tbl.iter().any(|(k, v)| {
-            if !NARROWING_KEYS.contains(&k.as_str()) {
-                return false;
-            }
-            // Treat `orphan = false` / `dead_end = false` as non-narrowing.
-            if matches!(k.as_str(), "orphan" | "dead_end" | "broken_links") {
-                return matches!(v, toml::Value::Boolean(true));
-            }
-            // List-typed narrowing keys with empty values don't narrow either.
-            if let toml::Value::Array(a) = v {
-                return !a.is_empty();
-            }
-            true
-        });
-
-        let has_fields = view_tbl.contains_key("fields");
-
-        if !has_narrowing && has_fields {
-            violations.push(Violation {
-                severity: Severity::Warn,
-                kind: None,
-                message: format!(
-                    "view '{name}' has no narrowing filter — `fields` controls display columns only, \
-                     not filtering. Did you mean `orphan = true` or `dead_end = true`?"
-                ),
-            });
-        } else if !has_narrowing {
-            violations.push(Violation {
-                severity: Severity::Warn,
-                kind: None,
-                message: format!(
-                    "view '{name}' has no narrowing filter — add at least one of: \
-                     tag, properties, task, orphan, dead_end, broken_links, glob, file, title"
-                ),
-            });
-        }
-    }
-
-    if violations.is_empty() {
-        None
-    } else {
-        Some(FileLintResult {
-            file: ".hyalo.toml".to_string(),
-            violations,
-        })
-    }
-}
-
-/// Validate the `[schema]` block in `.hyalo.toml` in isolation and return a
-/// pseudo-file lint result when it fails to parse.
-///
-/// Review round finding 2: `parse_schema_from_toml` (`crates/hyalo-cli/src/config.rs`)
-/// already detects a malformed `[schema]` block (an unknown key from
-/// `deny_unknown_fields`, an invalid field combination, ...) and degrades
-/// gracefully to "no schema validation" — but the only signal was a
-/// `-q`-suppressible stderr warning. That meant one typo'd key in ANY type's
-/// property block silently disabled schema enforcement for the WHOLE vault,
-/// while `lint --strict` printed a clean "no issues" and exited 0 on a file
-/// with a real violation the (silently disabled) schema would have caught.
-///
-/// This independently re-parses `.hyalo.toml` (mirroring [`validate_views`]'s
-/// pattern of representing a config-level problem as a violation on a
-/// `.hyalo.toml` pseudo-file) via the same
-/// [`crate::config::try_parse_schema_from_toml`] used by the runtime config
-/// loader, so the error text — including which key or value is wrong — is
-/// identical to the one already sent to stderr; this just also makes it a
-/// visible lint-result violation. `strict` controls severity directly (not
-/// promoted later like the per-file `schema/*` kinds): `Error` under
-/// `--strict` so `lint --strict` exits non-zero, `Warn` otherwise so
-/// `--format json` output shows the problem in `results` without also
-/// tripping the plain-`lint` exit code — a stricter stance than the
-/// per-file warnings needs, since "validation is silently off" is worse than
-/// any single file's violation it might have caught.
-///
-/// Returns `None` when the schema parses cleanly (including "no `[schema]`
-/// block at all", which is valid — not every vault uses schema validation).
-pub fn validate_schema_config(dir: &Path, strict: bool) -> Option<FileLintResult> {
-    let toml_path = dir.join(".hyalo.toml");
-    let contents = std::fs::read_to_string(&toml_path).ok()?;
-    let table: toml::Table = toml::from_str(&contents).ok()?;
-    match crate::config::try_parse_schema_from_toml(table.get("schema")) {
-        Ok(_) => None,
-        Err(message) => Some(FileLintResult {
-            file: ".hyalo.toml".to_string(),
-            violations: vec![Violation {
-                severity: if strict {
-                    Severity::Error
-                } else {
-                    Severity::Warn
-                },
-                kind: Some(VIOLATION_KIND_SCHEMA_MALFORMED),
-                message,
-            }],
-        }),
-    }
-}
-
-/// Compute lint counts from pre-indexed `IndexEntry` properties.
-///
-/// Used by `hyalo summary` to avoid re-reading files from disk.
-/// The `index_entries` iterator yields `(rel_path, properties)` tuples.
-/// Build a [`globset::GlobSet`] from `[lint] ignore` patterns for path-based
-/// exclusion. Returns `None` when `patterns` is empty or every pattern failed
-/// to compile (callers treat `None` as "exclude nothing"). Invalid patterns are
-/// reported via [`crate::warn::warn`] so the behaviour matches `hyalo lint`,
-/// which uses the same matching rules (`literal_separator(true)`,
-/// `backslash_escape(true)`). Paths are matched against their vault-relative
-/// form with `/` separators.
-pub(crate) fn build_lint_ignore_globset(patterns: &[String]) -> Option<globset::GlobSet> {
-    use globset::{GlobBuilder, GlobSetBuilder};
-    if patterns.is_empty() {
-        return None;
-    }
-    let mut builder = GlobSetBuilder::new();
-    let mut any_ok = false;
-    for pat in patterns {
-        match GlobBuilder::new(pat)
-            .literal_separator(true)
-            .backslash_escape(true)
-            .build()
-        {
-            Ok(g) => {
-                builder.add(g);
-                any_ok = true;
-            }
-            Err(e) => {
-                crate::warn::warn(format!("invalid [lint] ignore pattern {pat:?}: {e}"));
-            }
-        }
-    }
-    if !any_ok {
-        return None;
-    }
-    builder.build().ok()
-}
-
-/// True when `rel_path` (vault-relative, any separator) is excluded by the
-/// `[lint] ignore` glob set. `None` set means "exclude nothing".
-pub(crate) fn is_lint_ignored(set: Option<&globset::GlobSet>, rel_path: &str) -> bool {
-    set.is_some_and(|s| s.is_match(rel_path.replace('\\', "/")))
-}
-
-// ---------------------------------------------------------------------------
-// Text formatter
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// Extended body-lint types (new output shape per plan)
-// ---------------------------------------------------------------------------
-
-/// A group of violations for one rule within one file.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct RuleGroup {
-    pub rule: String,
-    pub count: usize,
-    pub shown: usize,
-    pub truncated: bool,
-    pub severity: String,
-    pub autofixable: bool,
-    pub violations: Vec<BodyViolation>,
-}
-
-/// A single body violation.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct BodyViolation {
-    pub line: usize,
-    pub column: usize,
-    /// Per-violation severity (`"error"` / `"warn"`). Carried alongside the
-    /// group severity because a folded group (notably `SCHEMA`) can mix the
-    /// two, and the text renderer must label each line with its own severity
-    /// so the display agrees with the `errors`/`warnings` counts (BUG-17).
-    pub severity: String,
-    pub message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub fix: Option<serde_json::Value>,
-}
-
-/// Extended lint output for one file (read-only shape).
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct ExtFileLintResult {
-    pub file: String,
-    /// Frontmatter `type:` discriminator, if the file declared one. Used by
-    /// the hint layer to surface `hyalo types show <T>` for SCHEMA failures.
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub doc_type: Option<String>,
-    /// Frontmatter + body violations grouped by rule.
-    pub rule_groups: Vec<RuleGroup>,
-}
-
-/// One entry in `fixed_groups`: a rule + count of violations that were fixed.
-/// Includes `violations` so text renderers can show line/message details.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct FixedGroup {
-    pub rule: String,
-    pub count: usize,
-    /// Violations that were fixed (same shape as `RuleGroup.violations`).
-    pub violations: Vec<BodyViolation>,
-}
-
-/// One entry in `conflicts`: a rule whose fix was skipped due to range overlap.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ConflictEntry {
-    pub rule: String,
-    pub reason: String,
-}
-
-/// Extended lint output for one file in fix-mode.
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct ExtFileLintFixResult {
-    pub file: String,
-    /// Frontmatter `type:` discriminator, if the file declared one. Mirrors
-    /// [`ExtFileLintResult::doc_type`] so the iter-143 SCHEMA-→-`types show`
-    /// hint also fires in `--fix` / `--fix --dry-run` output.
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub doc_type: Option<String>,
-    /// Rules that had fixes applied (or would be in DryRun).
-    pub fixed_groups: Vec<FixedGroup>,
-    /// Rules with violations that remain after fixing.
-    pub remaining_groups: Vec<RuleGroup>,
-    /// Rules whose fixes were skipped due to conflicts.
-    pub conflicts: Vec<ConflictEntry>,
-}
-
-/// Full extended lint output (read-only mode).
-///
-/// `Default` backs the empty-result shape `run.rs` emits when
-/// `--files-from` resolves to zero files — serializing `Self::default()`
-/// (with only `dry_run` overridden) keeps that shape from drifting out of
-/// sync with the real field set as fields are added, renamed or removed
-/// here, mirroring [`ExtLintFixOutput`]'s empty-result shape.
-#[derive(Debug, Default, serde::Serialize, serde::Deserialize)]
-pub struct ExtLintOutput {
-    pub files: Vec<ExtFileLintResult>,
-    /// Total number of violations found across all files.
-    ///
-    /// iter-216 D-1..D-5: named `violations`, not `total`. `results.total` is
-    /// reserved for "the number of items the command considered" (the
-    /// denominator); this is a count of findings, and the envelope's own
-    /// `total` on the very same document is the *file* count, so one name
-    /// used to carry two quantities in one payload.
-    pub violations: usize,
-    pub rules_fired: usize,
-    pub files_with_violations: usize,
-    /// Total number of files that were examined (including clean files).
-    pub files_checked: usize,
-    pub files_truncated: bool,
-    /// Number of error-severity violations.
-    pub errors: usize,
-    /// Number of warn-severity violations.
-    pub warnings: usize,
-    /// Files dropped from this run by `[lint] ignore` (UX-1, dogfood pre3).
-    /// Always present (not skipped when zero) — consistent with the other
-    /// count fields on this struct.
-    pub files_ignored: usize,
-    /// `true` when `--fix --dry-run` previewed fixes without writing them.
-    ///
-    /// iter-216 D-4: always present, never skipped when `false`. Top-level
-    /// `results` keys are always present; `set`/`remove`/`append`/`mv` already
-    /// emit `dry_run: false`, and a script switching between those and `lint`
-    /// used to get `null` from one and `false` from the other.
-    pub dry_run: bool,
-    /// Frontmatter fix actions applied (or previewed) per file.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub fixes: Vec<FileFixResult>,
-}
-
-/// Full extended lint output in fix-mode.
-///
-/// `Default` backs the empty-result shape `run.rs` emits when
-/// `--files-from` resolves to zero files (review finding #5) — serializing
-/// `Self::default()` (with only `dry_run` overridden) keeps that shape from
-/// drifting out of sync with the real field set as fields are added or
-/// removed here. `dry_run` is always serialized (iter-216 D-4), so the
-/// empty shape carries it as a plain `false`/`true`, same as the non-empty
-/// path.
-#[derive(Debug, Default, serde::Serialize, serde::Deserialize)]
-pub struct ExtLintFixOutput {
-    pub files: Vec<ExtFileLintFixResult>,
-    pub total_fixed: usize,
-    pub total_remaining: usize,
-    pub total_conflicts: usize,
-    pub rules_fired: usize,
-    pub files_with_violations: usize,
-    pub files_checked: usize,
-    pub files_truncated: bool,
-    /// Error-severity violations left unfixed after this run.
-    ///
-    /// Named `remaining_errors`/`remaining_warnings` (not `errors`/
-    /// `warnings`, iter-218 NEW-6b) because those keys mean something
-    /// different on [`ExtLintOutput`]: whole-run severity counts, not a
-    /// remaining-after-fix count. The two commands used to share a key name
-    /// for two different quantities — a script reading `.errors` off both
-    /// `lint` and `lint --fix` JSON silently got answers to different
-    /// questions.
-    pub remaining_errors: usize,
-    pub remaining_warnings: usize,
-    /// `true` when `--dry-run` previewed fixes without writing them.
-    /// Always present (iter-216 D-4) — see [`ExtLintOutput::dry_run`].
-    pub dry_run: bool,
-}
-
-/// Options for the extended lint run.
-pub struct ExtLintOptions<'a> {
-    pub fix: FixMode,
-    pub detailed: bool,
-    pub rule_filter: Option<&'a str>,
-    pub rule_prefix: Option<&'a str>,
-    pub max_per_rule: usize,
-    pub max_files: usize,
-    pub fix_rules: &'a [String],
-    /// Snapshot index for patching after fixes.
-    pub snapshot_index: &'a mut Option<hyalo_core::index::SnapshotIndex>,
-    pub index_path: Option<&'a Path>,
-    pub vault_dir: &'a Path,
-    /// When `true`, promote "no 'type' property" and "undeclared property in
-    /// frontmatter" from `Severity::Warn` to `Severity::Error`.
-    pub strict: bool,
-    /// When `true`, run the OKF conformance profile's advisory (warn-level)
-    /// rules in addition to the schema pass. Set by `hyalo lint --profile okf`.
-    pub okf_profile: bool,
-    /// When `true`, run the MADR conformance profile's advisory (warn-level)
-    /// rules in addition to the schema pass. Set by `hyalo lint --profile madr`.
-    pub madr_profile: bool,
-    /// When `true`, run the Agent Skills conformance profile's rules
-    /// (mostly warn-level; `SKILL-RESERVED-NAME` defaults to error) in
-    /// addition to the schema pass. Set by `hyalo lint --profile skills`.
-    pub skills_profile: bool,
-    /// When `true`, run the Keep a Changelog conformance profile's rules
-    /// (mostly error-level grammar/ordering; empty-section and link-ref
-    /// cross-check default to warn) in addition to the schema pass. Set by
-    /// `hyalo lint --profile changelog`.
-    pub changelog_profile: bool,
-    /// Resolved `[links] case_insensitive` mode for `vault_dir` (see
-    /// [`hyalo_core::mode_enabled`]). Used so `[schema] exempt`
-    /// globs (e.g. `**/index.md`) fold case the same way `hyalo okf index`
-    /// does on case-insensitive filesystems.
-    pub case_insensitive: bool,
-    /// Vault-wide context for the HYALO006 broken-link rule, built ONCE per
-    /// invocation in the dispatch arm (link graph / case index) and shared by
-    /// reference across the rayon workers. `None` when HYALO006 is disabled or
-    /// filtered out, so no graph is built and the rule never runs.
-    pub link_lint_ctx: Option<hyalo_mdlint::profiles::link::LinkLintContext>,
-    /// Files dropped from this run by `[lint] ignore`, regardless of scope
-    /// (bare sweep, `--glob`, or `--file`). UX-1 (dogfood pre3): surfaced in
-    /// the bare-sweep summary line so a full-vault run stops silently hiding
-    /// how much of the vault it skipped (`68 files checked (318 ignored by
-    /// [lint] ignore)`) — the named-file/glob-all-ignored cases already get a
-    /// dedicated stderr notice from the caller.
-    pub files_ignored: usize,
-}
-
-/// Run the extended lint (frontmatter + body) and return the new output shape.
-#[allow(clippy::too_many_arguments)]
-pub fn lint_files_extended(
-    files: &[(std::path::PathBuf, String)],
-    schema: &SchemaConfig,
-    md_lint_engine: &hyalo_mdlint::HyaloLintEngine,
-    md_lint_config: &hyalo_mdlint::LintConfig,
-    opts: &mut ExtLintOptions<'_>,
-) -> Result<(CommandOutcome, LintCounts)> {
-    #[cfg(not(miri))]
-    use rayon::prelude::*;
-
-    // Build rule filter list
-    let rule_filter: Vec<String> = match (opts.rule_filter, opts.rule_prefix) {
-        (Some(rule), _) => vec![rule.to_owned()],
-        // M-10: prefix matching is case-insensitive so `--rule-prefix hyalo`
-        // selects the same family as `--rule-prefix HYALO`.
-        (None, Some(prefix)) => md_lint_engine
-            .rules_matching_prefix_ci(prefix)
-            .iter()
-            .map(|e| e.id.clone())
-            .collect(),
-        (None, None) => vec![],
-    };
-
-    // Determine if schema has `status: completed` in any type.
-    let schema_has_completed = schema_has_completed_status(schema);
-
-    let strict = opts.strict;
-    let okf_profile = opts.okf_profile;
-    let madr_profile = opts.madr_profile;
-    let skills_profile = opts.skills_profile;
-    let changelog_profile = opts.changelog_profile;
-    let vault_dir = opts.vault_dir;
-    let case_insensitive = opts.case_insensitive;
-    // Shared HYALO006 context (built once in dispatch); borrowed by every
-    // worker. `None` disables the rule for this run.
-    let link_ctx = opts.link_lint_ctx.as_ref();
-
-    // Process files in parallel. Each worker lints one file.
-    let lint_file = |(full_path, rel_path): &(std::path::PathBuf, String)| {
-        lint_one_file_extended(
-            full_path,
-            rel_path,
-            schema,
-            md_lint_engine,
-            md_lint_config,
-            &rule_filter,
-            schema_has_completed,
-            opts.fix,
-            opts.fix_rules,
-            opts.max_per_rule,
-            strict,
-            okf_profile,
-            madr_profile,
-            skills_profile,
-            changelog_profile,
-            vault_dir,
-            case_insensitive,
-            link_ctx,
-        )
-    };
-    #[cfg(not(miri))]
-    let per_file: Vec<Result<PerFileLintResult>> = files.par_iter().map(lint_file).collect();
-    #[cfg(miri)]
-    let per_file: Vec<Result<PerFileLintResult>> = files.iter().map(lint_file).collect();
-
-    // Merge results serially.
-    let mut all_results: Vec<PerFileLintResult> = Vec::with_capacity(files.len());
-    let mut modified_files: Vec<String> = Vec::new();
-
-    for result in per_file {
-        let mut r = result?;
-        if r.body_modified {
-            modified_files.push(r.rel_path.clone());
-            r.body_modified = false;
-        }
-        all_results.push(r);
-    }
-
-    // Patch index for body-modified files (ARCH-3, iter-226): through the
-    // MutationJournal so entry AND link graph stay current, flushed once.
-    if !modified_files.is_empty() {
-        let mut journal =
-            crate::commands::journal::MutationJournal::new(opts.snapshot_index, opts.index_path);
-        journal.rescan_modified(opts.vault_dir, &modified_files)?;
-        journal.flush()?;
-    }
-
-    let is_fix_mode = matches!(opts.fix, FixMode::Apply | FixMode::DryRun);
-
-    // Sort by total violations descending (worst offenders first).
-    // UX-4 (dogfood v0.20.0): error-carrying files sort ahead of
-    // warning-only files (then by total violations) — the display cap
-    // drops the least severe files first, so "4 errors" hidden behind a
-    // 50-file listing of pure warnings can no longer happen.
-    all_results.sort_by_key(|r| {
-        let (errors, _) = count_file_errors_warnings(r, is_fix_mode);
-        (
-            std::cmp::Reverse(errors),
-            std::cmp::Reverse(r.total_violations),
-        )
-    });
-
-    // Cap files.
-    let total_files_with_violations = all_results
-        .iter()
-        .filter(|r| r.total_violations > 0)
-        .count();
-    let files_checked_total = all_results.len();
-    // `files_truncated` describes the *listed* files, not the examined ones.
-    // `all_results` is sorted worst-first, so the display cap only ever drops a
-    // violating file once there are more of those than `max_files`. Deriving it
-    // from `files_checked > max_files` (as this did before iter-210 BUG-6)
-    // false-positived on every clean vault larger than the limit.
-    let files_truncated = total_files_with_violations > opts.max_files;
-
-    // Authoritative error/warning totals, computed over EVERY result *before*
-    // the display list is capped to `max_files`. The per-file display loops
-    // below only build the (possibly truncated) `files[]` array — they must not
-    // be the source of the summary counters, or a `--limit`/`max_files` cap
-    // would under-report errors and let a corrupt vault exit 0 (ff-rdp B5).
-    let (authoritative_errors, authoritative_warnings) =
-        count_errors_warnings(&all_results, is_fix_mode);
-
-    // Whole-run violation total and distinct firing rules, computed over EVERY
-    // result *before* the display cap, exactly like the error/warning totals
-    // above. Deriving them from the capped display loops (pre-iter-210) made
-    // JSON `total` describe a different run than `errors` + `warnings` — on a
-    // large vault the two disagreed by an order of magnitude (BUG-6).
-    let authoritative_total: usize = all_results.iter().map(|r| r.total_violations).sum();
-    let authoritative_rules_fired = {
-        let mut seen: std::collections::HashSet<&str> = std::collections::HashSet::new();
-        for r in &all_results {
-            for rule_id in r.violations_by_rule.keys() {
-                seen.insert(rule_id.as_str());
-            }
-        }
-        seen.len()
-    };
-
-    // Fix-mode totals, computed over EVERY result *before* the display cap,
-    // exactly like `authoritative_total`/`authoritative_rules_fired` above.
-    // The per-file display loop below only builds the (possibly truncated)
-    // `files[]` array for fix-mode too — deriving `total_fixed` /
-    // `total_remaining` / `total_conflicts` from it (as this did before
-    // iter-218) made `--limit` silently understate `conflicts`, hiding
-    // apply-time surprises from anyone who dry-ran at the default limit
-    // (NEW-6).
-    let (authoritative_total_fixed, authoritative_total_remaining, authoritative_total_conflicts) =
-        if is_fix_mode {
-            all_results
-                .iter()
-                .fold((0, 0, 0), |(fixed, remaining, conflicts), r| {
-                    let (f, rem, c) = fix_mode_file_totals(r);
-                    (fixed + f, remaining + rem, conflicts + c)
-                })
-        } else {
-            (0, 0, 0)
-        };
-
-    // Cap the display list (0 == unlimited is normalized to `usize::MAX` by the
-    // caller, so truncation here is a no-op for `--limit 0`).
-    all_results.truncate(opts.max_files);
-
-    let val = if is_fix_mode {
-        // -------------------------------------------------------------------
-        // Fix-mode output: fixed_groups / remaining_groups / conflicts shape.
-        // -------------------------------------------------------------------
-        let mut output_fix_files: Vec<ExtFileLintFixResult> = Vec::new();
-
-        for r in &all_results {
-            // Build a set of (rule_id, order-within-rule) → FixOutcome from
-            // body_fix_outcomes.  We use a Vec<(rule, outcome)> indexed by the
-            // position among fixable diagnostics (same order as fixable_indices).
-            // Easier to just build maps indexed by rule.
-            let mut applied_by_rule: indexmap::IndexMap<String, Vec<BodyViolation>> =
-                indexmap::IndexMap::new();
-            let mut conflict_by_rule: indexmap::IndexMap<String, String> =
-                indexmap::IndexMap::new();
-
-            for (rule_id, outcome) in &r.body_fix_outcomes {
-                match outcome {
-                    FixOutcome::Applied => {
-                        applied_by_rule.entry(rule_id.clone()).or_default();
-                    }
-                    FixOutcome::Conflict { blocking_rule } => {
-                        conflict_by_rule
-                            .entry(rule_id.clone())
-                            .or_insert_with(|| blocking_rule.clone());
-                    }
-                    FixOutcome::NoFix => {}
-                }
-            }
-
-            // SCHEMA fixed count: derived from re-validating post-fix
-            // properties (resolved = before - after). Falls back to 0 when
-            // the post-fix re-validation didn't run.
-            let schema_before = r
-                .violations_by_rule
-                .get("SCHEMA")
-                .map_or(0, std::vec::Vec::len);
-            let schema_after = r
-                .post_fix_schema_remaining
-                .as_ref()
-                .map_or(schema_before, std::vec::Vec::len);
-            let schema_fix_count = schema_before.saturating_sub(schema_after);
-
-            // fixed_groups: rules with at least one applied fix + SCHEMA if fixed.
-            let mut fixed_groups: Vec<FixedGroup> = Vec::new();
-            if schema_fix_count > 0 {
-                fixed_groups.push(FixedGroup {
-                    rule: "SCHEMA".to_owned(),
-                    count: schema_fix_count,
-                    violations: Vec::new(),
-                });
-            }
-            for (rule_id, _) in &applied_by_rule {
-                // Surface only diagnostics whose fix was actually Applied —
-                // not the entire `violations_by_rule[rule_id]` set, which
-                // also contains conflicts and not-selected entries for the
-                // same rule.
-                let viols: Vec<BodyViolation> = r
-                    .violations_by_rule
-                    .get(rule_id)
-                    .map(|vs| {
-                        vs.iter()
-                            .filter(|v| v.fixed)
-                            .take(opts.max_per_rule)
-                            .map(|v| BodyViolation {
-                                line: v.line,
-                                column: v.column,
-                                severity: v.severity.clone(),
-                                message: v.message.clone(),
-                                fix: v.fix.clone(),
-                            })
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                let count = r
-                    .violations_by_rule
-                    .get(rule_id)
-                    .map_or(0, |vs| vs.iter().filter(|v| v.fixed).count());
-                if count > 0 {
-                    fixed_groups.push(FixedGroup {
-                        rule: rule_id.clone(),
-                        count,
-                        violations: viols,
-                    });
-                }
-            }
-
-            // remaining_groups: violations not fixed.
-            let mut remaining_groups: Vec<RuleGroup> = Vec::new();
-            for (rule_id, violations) in &r.violations_by_rule {
-                // SCHEMA: post-fix remaining set comes from re-running
-                // `validate_properties` against the mutated frontmatter,
-                // not from a fix_actions count heuristic.
-                if rule_id == "SCHEMA" {
-                    let remaining_owned: Vec<InternalViolation> =
-                        if let Some(post) = r.post_fix_schema_remaining.as_ref() {
-                            post.iter()
-                                .map(|v| InternalViolation {
-                                    line: v.line,
-                                    column: v.column,
-                                    message: v.message.clone(),
-                                    severity: v.severity.clone(),
-                                    fix: v.fix.clone(),
-                                    fixed: v.fixed,
-                                    autofixable: v.autofixable,
-                                })
-                                .collect()
-                        } else {
-                            // No fix-mode SCHEMA pass ran (e.g., --fix-rule
-                            // filtered SCHEMA out). All originals remain.
-                            violations
-                                .iter()
-                                .map(|v| InternalViolation {
-                                    line: v.line,
-                                    column: v.column,
-                                    message: v.message.clone(),
-                                    severity: v.severity.clone(),
-                                    fix: v.fix.clone(),
-                                    fixed: v.fixed,
-                                    autofixable: v.autofixable,
-                                })
-                                .collect()
-                        };
-                    let remaining = remaining_owned.len();
-                    if remaining == 0 {
-                        continue;
-                    }
-                    let shown = remaining.min(opts.max_per_rule);
-                    let truncated = remaining > shown;
-                    let body_violations = remaining_owned
-                        .iter()
-                        .take(shown)
-                        .map(|v| BodyViolation {
-                            line: v.line,
-                            column: v.column,
-                            severity: v.severity.clone(),
-                            message: v.message.clone(),
-                            fix: v.fix.clone(),
-                        })
-                        .collect();
-                    remaining_groups.push(RuleGroup {
-                        rule: rule_id.clone(),
-                        count: remaining,
-                        shown,
-                        truncated,
-                        severity: group_severity(&remaining_owned),
-                        // Autofixable only if at least one remaining SCHEMA
-                        // violation could still be fixed (mapl BUG-3).
-                        autofixable: remaining_owned
-                            .iter()
-                            .any(|v| v.autofixable.unwrap_or(true)),
-                        violations: body_violations,
-                    });
-                    continue;
-                }
-
-                // Body rules: filter out violations that were actually fixed.
-                let remaining_violations: Vec<&InternalViolation> =
-                    violations.iter().filter(|v| !v.fixed).collect();
-                let remaining_count = remaining_violations.len();
-                if remaining_count == 0 {
-                    continue;
-                }
-
-                let autofixable = md_lint_engine
-                    .available_rules()
-                    .iter()
-                    .find(|e| &e.id == rule_id)
-                    .is_some_and(|e| e.autofixable);
-                let severity = remaining_violations
-                    .first()
-                    .map_or_else(|| "warn".to_owned(), |v| v.severity.clone());
-                let shown = remaining_count.min(opts.max_per_rule);
-                let truncated = remaining_count > shown;
-                let body_violations = remaining_violations
-                    .iter()
-                    .take(shown)
-                    .map(|v| BodyViolation {
-                        line: v.line,
-                        column: v.column,
-                        severity: v.severity.clone(),
-                        message: v.message.clone(),
-                        fix: v.fix.clone(),
-                    })
-                    .collect();
-                remaining_groups.push(RuleGroup {
-                    rule: rule_id.clone(),
-                    count: remaining_count,
-                    shown,
-                    truncated,
-                    severity,
-                    autofixable,
-                    violations: body_violations,
-                });
-            }
-            remaining_groups.sort_by_key(|g| std::cmp::Reverse(g.count));
-
-            // conflicts: rules with at least one conflicting fix.
-            let mut conflicts: Vec<ConflictEntry> = Vec::new();
-            for (rule_id, blocking_rule) in &conflict_by_rule {
-                conflicts.push(ConflictEntry {
-                    rule: rule_id.clone(),
-                    reason: format!("range overlap with {blocking_rule}"),
-                });
-            }
-
-            if !fixed_groups.is_empty() || !remaining_groups.is_empty() || !conflicts.is_empty() {
-                output_fix_files.push(ExtFileLintFixResult {
-                    file: r.rel_path.clone(),
-                    doc_type: r.doc_type.clone(),
-                    fixed_groups,
-                    remaining_groups,
-                    conflicts,
-                });
-            }
-        }
-
-        let fix_output = ExtLintFixOutput {
-            files: output_fix_files,
-            total_fixed: authoritative_total_fixed,
-            total_remaining: authoritative_total_remaining,
-            total_conflicts: authoritative_total_conflicts,
-            rules_fired: authoritative_rules_fired,
-            files_with_violations: total_files_with_violations,
-            files_checked: files_checked_total,
-            files_truncated,
-            remaining_errors: authoritative_errors,
-            remaining_warnings: authoritative_warnings,
-            dry_run: matches!(opts.fix, FixMode::DryRun),
-        };
-        serde_json::to_value(&fix_output).context("failed to serialize fix lint output")?
-    } else {
-        // -------------------------------------------------------------------
-        // Read-only output: unchanged rule_groups shape.
-        // -------------------------------------------------------------------
-        let mut output_files: Vec<ExtFileLintResult> = Vec::new();
-        let mut all_fix_actions: Vec<FileFixResult> = Vec::new();
-
-        for r in &all_results {
-            if !r.fix_actions.is_empty() {
-                all_fix_actions.push(FileFixResult {
-                    file: r.rel_path.clone(),
-                    actions: r.fix_actions.clone(),
-                });
-            }
-            if r.total_violations == 0 {
-                continue;
-            }
-            let mut rule_groups: Vec<RuleGroup> = Vec::new();
-            for (rule_id, violations) in &r.violations_by_rule {
-                let count = violations.len();
-
-                let autofixable = if rule_id == "SCHEMA" {
-                    // The SCHEMA group folds several checks; it is autofixable
-                    // only if at least one member is (per-violation
-                    // `autofixable`). A group made up entirely of
-                    // non-synthesizable missing-required violations reports
-                    // `false` (mapl BUG-3).
-                    violations.iter().any(|v| v.autofixable.unwrap_or(true))
-                } else {
-                    md_lint_engine
-                        .available_rules()
-                        .iter()
-                        .find(|e| &e.id == rule_id)
-                        .is_some_and(|e| e.autofixable)
-                };
-                let severity = group_severity(violations);
-
-                let shown = if opts.detailed {
-                    violations.len()
-                } else {
-                    violations.len().min(opts.max_per_rule)
-                };
-                let truncated = count > shown;
-                let body_violations: Vec<BodyViolation> = violations[..shown]
-                    .iter()
-                    .map(|v| BodyViolation {
-                        line: v.line,
-                        column: v.column,
-                        severity: v.severity.clone(),
-                        message: v.message.clone(),
-                        fix: v.fix.clone(),
-                    })
-                    .collect();
-
-                rule_groups.push(RuleGroup {
-                    rule: rule_id.clone(),
-                    count,
-                    shown,
-                    truncated,
-                    severity,
-                    autofixable,
-                    violations: body_violations,
-                });
-            }
-            rule_groups.sort_by_key(|g| std::cmp::Reverse(g.count));
-            output_files.push(ExtFileLintResult {
-                file: r.rel_path.clone(),
-                doc_type: r.doc_type.clone(),
-                rule_groups,
-            });
-        }
-
-        let output = ExtLintOutput {
-            files: output_files,
-            violations: authoritative_total,
-            rules_fired: authoritative_rules_fired,
-            files_with_violations: total_files_with_violations,
-            files_checked: files_checked_total,
-            files_truncated,
-            errors: authoritative_errors,
-            warnings: authoritative_warnings,
-            files_ignored: opts.files_ignored,
-            dry_run: false,
-            fixes: all_fix_actions,
-        };
-        serde_json::to_value(&output).context("failed to serialize extended lint output")?
-    };
-
-    // Counts drive the exit code and `hyalo summary`. Use the authoritative
-    // pre-truncation totals so a `--limit`/`max_files` cap can never mask an
-    // error (which would let a corrupt vault exit 0). `total_errors` /
-    // `total_warnings` accumulated by the display loops only ever cover the
-    // capped `files[]` slice and would under-report.
-    let counts = LintCounts {
-        errors: authoritative_errors,
-        warnings: authoritative_warnings,
-        files_with_issues: total_files_with_violations,
-    };
-
-    let outcome = CommandOutcome::success_with_total(
-        crate::output::format_success(Format::Json, &val),
-        total_files_with_violations as u64,
-    );
-
-    Ok((outcome, counts))
-}
-
-/// Per-file violation entry (internal).
-struct InternalViolation {
-    line: usize,
-    column: usize,
-    message: String,
-    severity: String,
-    fix: Option<serde_json::Value>,
-    /// True when this body diagnostic's fix was successfully applied during
-    /// the current fix-mode run. Always `false` for read-only and frontmatter
-    /// (SCHEMA) violations — frontmatter fixes are tracked separately via
-    /// `fix_actions`.
-    fixed: bool,
-    /// Whether `--fix` could resolve THIS specific violation. Meaningful for
-    /// SCHEMA violations, whose group otherwise reports a single coarse
-    /// `autofixable`: a "missing required property" with no declared default
-    /// cannot be synthesized (mapl BUG-3), so it is `false` here while a
-    /// missing-but-defaulted property is `true`. `None` = use the group/rule
-    /// default (body rules, where autofixability is a property of the rule).
-    autofixable: Option<bool>,
-}
-
-/// Severity label for a group of violations under one rule id.
-///
-/// Most rule groups hold violations that all share one severity (hyalo
-/// assigns severity per rule, not per violation), but the synthetic
-/// `"SCHEMA"` group folds together several distinct checks (missing
-/// required field, undeclared property, missing type, ...) that can mix
-/// error and warn — using `violations.first()` there mislabels the whole
-/// group whenever a warn happens to be first. Returns the max severity
-/// across the group instead, which is correct for both the uniform and
-/// mixed cases.
-fn group_severity(violations: &[InternalViolation]) -> String {
-    if violations.iter().any(|v| v.severity == "error") {
-        "error".to_owned()
-    } else {
-        "warn".to_owned()
-    }
-}
-
-/// Outcome for a single diagnostic's fix attempt (internal, used in fix-mode).
-#[derive(Debug)]
-enum FixOutcome {
-    /// Fix was applied (or would be in DryRun).
-    Applied,
-    /// Fix conflicted with another fix's range.
-    Conflict { blocking_rule: String },
-    /// Diagnostic had no fix, or fix was not selected.
-    NoFix,
-}
-
-/// Per-file lint result (internal, before grouping).
-struct PerFileLintResult {
-    rel_path: String,
-    /// Frontmatter `type:` discriminator, if declared. Propagated into
-    /// `ExtFileLintResult.doc_type` for the hint layer.
-    doc_type: Option<String>,
-    violations_by_rule: indexmap::IndexMap<String, Vec<InternalViolation>>,
-    total_violations: usize,
-    body_modified: bool,
-    /// Frontmatter fix actions applied or previewed.
-    fix_actions: Vec<FixAction>,
-    /// Fix outcomes keyed by (rule_id, index_within_rule).
-    /// Only populated in fix-mode.
-    body_fix_outcomes: Vec<(String, FixOutcome)>,
-    /// SCHEMA (frontmatter) violations remaining *after* applying fixes,
-    /// computed by re-running `validate_properties` against the mutated
-    /// frontmatter. `Some(vec![])` means all SCHEMA violations were resolved;
-    /// `None` means fix-mode was off or no SCHEMA pass ran. Body rules use
-    /// `InternalViolation.fixed` instead.
-    post_fix_schema_remaining: Option<Vec<InternalViolation>>,
-}
-
-/// Compute one file's fix-mode totals — (fixed, remaining, conflicts) —
-/// without building the display groups (`fixed_groups`/`remaining_groups`/
-/// `conflicts`). Mirrors the per-file accounting the display loop in
-/// [`run_ext_lint`] does inline, but is called separately over the *full*
-/// `all_results` before the `--limit` display cap is applied, so
-/// `total_fixed`/`total_remaining`/`total_conflicts` describe the whole run
-/// even when the listing is truncated (iter-218 NEW-6).
-fn fix_mode_file_totals(r: &PerFileLintResult) -> (usize, usize, usize) {
-    let mut applied_rules: HashSet<&str> = HashSet::new();
-    let mut conflict_rules: HashSet<&str> = HashSet::new();
-    for (rule_id, outcome) in &r.body_fix_outcomes {
-        match outcome {
-            FixOutcome::Applied => {
-                applied_rules.insert(rule_id.as_str());
-            }
-            FixOutcome::Conflict { .. } => {
-                conflict_rules.insert(rule_id.as_str());
-            }
-            FixOutcome::NoFix => {}
-        }
-    }
-
-    let schema_before = r
-        .violations_by_rule
-        .get("SCHEMA")
-        .map_or(0, std::vec::Vec::len);
-    let schema_after = r
-        .post_fix_schema_remaining
-        .as_ref()
-        .map_or(schema_before, std::vec::Vec::len);
-    let mut fixed = schema_before.saturating_sub(schema_after);
-    let mut remaining = 0usize;
-
-    for (rule_id, violations) in &r.violations_by_rule {
-        if rule_id == "SCHEMA" {
-            remaining += schema_after;
-            continue;
-        }
-        if applied_rules.contains(rule_id.as_str()) {
-            fixed += violations.iter().filter(|v| v.fixed).count();
-        }
-        remaining += violations.iter().filter(|v| !v.fixed).count();
-    }
-
-    (fixed, remaining, conflict_rules.len())
-}
-
-/// Count the error- and warning-severity violations across every result,
-/// independent of the `max_files` display cap.
-///
-/// In read-only mode this is simply every violation, reported as
-/// [`ExtLintOutput::errors`]/`::warnings`. In fix mode it counts the
-/// violations that *remain* after fixing — SCHEMA remainders come from
-/// `post_fix_schema_remaining`, body remainders from `!v.fixed` — reported as
-/// [`ExtLintFixOutput::remaining_errors`]/`::remaining_warnings` (renamed
-/// from `errors`/`warnings` in iter-218 NEW-6b, since this mode's count means
-/// something different from the read-only mode's despite the pre-rename JSON
-/// using the same key for both). Either way, the exit code they drive
-/// reflects the true whole-vault state, not just the first `max_files`
-/// shown.
-fn count_errors_warnings(results: &[PerFileLintResult], is_fix_mode: bool) -> (usize, usize) {
-    results
-        .iter()
-        .map(|r| count_file_errors_warnings(r, is_fix_mode))
-        .fold((0, 0), |(e, w), (fe, fw)| (e + fe, w + fw))
-}
-
-/// One file's contribution to [`count_errors_warnings`] — (errors, warnings).
-///
-/// Also used by the display ordering (UX-4, dogfood v0.20.0): files that
-/// carry *errors* sort ahead of warning-only files so a display cap can
-/// never push the run's errors out of the listed slice.
-fn count_file_errors_warnings(r: &PerFileLintResult, is_fix_mode: bool) -> (usize, usize) {
-    let mut errors = 0usize;
-    let mut warnings = 0usize;
-    let mut tally = |severity: &str| {
-        if severity == "error" {
-            errors += 1;
-        } else {
-            warnings += 1;
-        }
-    };
-    for (rule_id, violations) in &r.violations_by_rule {
-        if is_fix_mode {
-            if rule_id == "SCHEMA" {
-                // Post-fix SCHEMA remainder, or all originals if no fix pass ran.
-                let remaining = r.post_fix_schema_remaining.as_ref().unwrap_or(violations);
-                for v in remaining {
-                    tally(&v.severity);
-                }
-            } else {
-                for v in violations.iter().filter(|v| !v.fixed) {
-                    tally(&v.severity);
-                }
-            }
-        } else {
-            for v in violations {
-                tally(&v.severity);
-            }
-        }
-    }
-    (errors, warnings)
-}
-
-/// Record a `--fix` write-path failure (TOCTOU or I/O) as a `FILE`-rule
-/// error violation, so it shows up in the report instead of only aborting
-/// via `?` (Finding 4, review round on PR #254 — the write-path sibling of
-/// M-1's read-path fix).
-fn push_fix_write_error_violation(
-    violations_by_rule: &mut indexmap::IndexMap<String, Vec<InternalViolation>>,
-    message: &str,
-) {
-    violations_by_rule
-        .entry("FILE".to_owned())
-        .or_default()
-        .push(InternalViolation {
-            line: 1,
-            column: 1,
-            message: message.to_owned(),
-            severity: "error".to_owned(),
-            fix: None,
-            fixed: false,
-            autofixable: None,
-        });
-}
-
-/// Lint a single file (frontmatter + body). Returns a `PerFileLintResult`.
-#[allow(clippy::too_many_arguments, clippy::fn_params_excessive_bools)]
-fn lint_one_file_extended(
-    full_path: &Path,
-    rel_path: &str,
-    schema: &SchemaConfig,
-    engine: &hyalo_mdlint::HyaloLintEngine,
-    md_lint_config: &hyalo_mdlint::LintConfig,
-    rule_filter: &[String],
-    schema_has_completed: bool,
-    fix: FixMode,
-    fix_rules: &[String],
-    max_per_rule: usize,
-    strict: bool,
-    okf_profile: bool,
-    madr_profile: bool,
-    skills_profile: bool,
-    changelog_profile: bool,
-    vault_dir: &Path,
-    case_insensitive: bool,
-    link_ctx: Option<&hyalo_mdlint::profiles::link::LinkLintContext>,
-) -> Result<PerFileLintResult> {
-    // One rule's fix can expose a fresh violation for another rule (e.g. a
-    // trimmed line changing what counts as a duplicate blank line), so a
-    // single lint→fix pass over the body does not always converge. Bounds
-    // the lint→fix→re-lint loop below.
-    const MAX_BODY_FIX_PASSES: usize = 5;
-
-    // Stat before reading: oversized files are skipped rather than loaded
-    // whole into memory (mirrors `scanner::scan_file_multi`'s own guard).
-    let meta =
-        std::fs::metadata(full_path).with_context(|| format!("failed to stat {rel_path}"))?;
-    if meta.len() > scanner::MAX_FILE_SIZE {
-        eprintln!(
-            "warning: skipping {} ({} MiB exceeds {} MiB limit)",
-            full_path.display(),
-            meta.len() / (1024 * 1024),
-            scanner::MAX_FILE_SIZE / (1024 * 1024)
-        );
-        let mut violations_by_rule = indexmap::IndexMap::new();
-        violations_by_rule.insert(
-            "FILE".to_owned(),
-            vec![InternalViolation {
-                line: 1,
-                column: 1,
-                message: format!(
-                    "file exceeds {} MiB size limit — skipped, not linted",
-                    scanner::MAX_FILE_SIZE / (1024 * 1024)
-                ),
-                severity: "warn".to_owned(),
-                fix: None,
-                fixed: false,
-                autofixable: None,
-            }],
-        );
-        return Ok(PerFileLintResult {
-            rel_path: rel_path.to_owned(),
-            doc_type: None,
-            violations_by_rule,
-            total_violations: 1,
-            body_modified: false,
-            fix_actions: Vec::new(),
-            body_fix_outcomes: Vec::new(),
-            post_fix_schema_remaining: None,
-        });
-    }
-
-    // Baseline mtime fingerprint for TOCTOU detection around fix-mode
-    // writes below. Derived from the stat above instead of a second
-    // `read_mtime` round-trip.
-    let mut mtime0: (std::time::SystemTime, u64) = (
-        meta.modified()
-            .with_context(|| format!("mtime not available for {rel_path}"))?,
-        meta.len(),
-    );
-
-    // Read the file content once. A single unreadable file (invalid UTF-8,
-    // permission error, etc.) must not abort the whole vault-wide run — the
-    // caller's merge loop propagates any `Err` here via `?`, which used to
-    // kill `lint`/`lint --fix` entirely on one corrupt file (M-1,
-    // adversarial-review-2026-08-23.md). Report it once and skip just this
-    // file, mirroring the size-limit skip above and the lossy-decode
-    // skip+warn precedent in `scanner/mod.rs`.
-    let content = match std::fs::read_to_string(full_path) {
-        Ok(content) => content,
-        Err(e) => {
-            eprintln!("warning: skipping {} ({e})", full_path.display());
-            let mut violations_by_rule = indexmap::IndexMap::new();
-            violations_by_rule.insert(
-                "FILE".to_owned(),
-                vec![InternalViolation {
-                    line: 1,
-                    column: 1,
-                    message: format!("could not read file ({e}) — skipped, not linted"),
-                    severity: "error".to_owned(),
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                }],
-            );
-            return Ok(PerFileLintResult {
-                rel_path: rel_path.to_owned(),
-                doc_type: None,
-                violations_by_rule,
-                total_violations: 1,
-                body_modified: false,
-                fix_actions: Vec::new(),
-                body_fix_outcomes: Vec::new(),
-                post_fix_schema_remaining: None,
-            });
-        }
-    };
-
-    // Find where the frontmatter ends so we can split body.
-    let body_start = find_body_start(&content);
-    let body_content = &content[body_start..];
-
-    // Frontmatter pass: use existing logic but convert to new shape.
-    let properties = match hyalo_core::frontmatter::read_frontmatter(full_path) {
-        Ok(p) => p,
-        Err(e) if hyalo_core::frontmatter::is_parse_error(&e) => {
-            // Malformed frontmatter — report as a single error-severity
-            // violation under the stable HYALO005 rule id so it shows up in
-            // lint output, `lint-rules list`, and CI. A file whose frontmatter
-            // cannot be parsed is otherwise invisible to every other rule, so
-            // this must never silently vanish (RB-3 / df-own-kb B3).
-            //
-            // Severity is user-configurable via `[lint.rules.HYALO005]
-            // severity = "warn"`, but the rule always emits (it is not
-            // disable-able) and no profile can downgrade it.
-            let severity = md_lint_config
-                .rules
-                .get(RULE_ID_FRONTMATTER_PARSE_ERROR)
-                .and_then(hyalo_mdlint::RuleOverride::severity)
-                .filter(|s| s.eq_ignore_ascii_case("warn"))
-                .map_or("error", |_| "warn");
-            let mut violations_by_rule = indexmap::IndexMap::new();
-            violations_by_rule.insert(
-                RULE_ID_FRONTMATTER_PARSE_ERROR.to_owned(),
-                vec![InternalViolation {
-                    line: 1,
-                    column: 1,
-                    message: format!(
-                        "{}: {}",
-                        crate::hints::PARSE_ERROR_PREFIX,
-                        terse_root_cause(&e)
-                    ),
-                    severity: severity.to_owned(),
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                }],
-            );
-            return Ok(PerFileLintResult {
-                rel_path: rel_path.to_owned(),
-                doc_type: None,
-                violations_by_rule,
-                total_violations: 1,
-                body_modified: false,
-                fix_actions: Vec::new(),
-                body_fix_outcomes: Vec::new(),
-                post_fix_schema_remaining: None,
-            });
-        }
-        Err(e) => return Err(e).context(format!("reading frontmatter from {rel_path}")),
-    };
-
-    let mut violations_by_rule: indexmap::IndexMap<String, Vec<InternalViolation>> =
-        indexmap::IndexMap::new();
-    // Set when the frontmatter fix write fails (TOCTOU or I/O) below, so the
-    // body-fix write is skipped for this file rather than attempted against
-    // a `mtime0` baseline that may no longer be trustworthy (Finding 4,
-    // review round on PR #254).
-    let mut frontmatter_write_failed = false;
-
-    // Frontmatter violations → use the existing `validate_properties` but map to new shape.
-    // Only emit if the rule isn't filtered out.
-    let should_include_frontmatter = rule_filter.is_empty()
-        || rule_filter
-            .iter()
-            .any(|r| r.starts_with("FRONTMATTER") || r == "SCHEMA");
-    if should_include_frontmatter {
-        let mut fm_violations =
-            validate_properties(rel_path, &properties, schema, case_insensitive);
-
-        // Under --strict, the missing-type warning is promoted to an error.
-        // `validate_properties` only emits it when the schema is non-empty, but
-        // strict mode should catch missing `type` regardless — inject it when
-        // the schema was empty and the property is absent (BUG-3 / iter-133).
-        let already_has_missing_type = fm_violations
-            .iter()
-            .any(|v| v.kind == Some(VIOLATION_KIND_MISSING_TYPE));
-        // Exempt / reserved files are bound to no schema, so `--strict` must
-        // not inject a missing-`type` error for them either. A path-bound file
-        // (bind = typing) is fully typed by its binding, so it is likewise
-        // exempt from the missing-`type` injection.
-        if strict
-            && !already_has_missing_type
-            && !properties.contains_key("type")
-            && !schema.exempt.is_exempt_ci(rel_path, case_insensitive)
-            && schema.bound_type_for(rel_path).is_none()
-        {
-            fm_violations.push(Violation {
-                severity: Severity::Warn,
-                kind: Some(VIOLATION_KIND_MISSING_TYPE),
-                message: "no 'type' property — validating against default schema only".to_owned(),
-            });
-        }
-
-        for v in fm_violations {
-            // In strict mode, promote the two targeted schema warnings to errors.
-            // Match on the stable `kind` identifier rather than message text so
-            // future message rewordings don't silently disable promotion.
-            let effective_severity = if strict
-                && v.severity == Severity::Warn
-                && matches!(
-                    v.kind,
-                    Some(VIOLATION_KIND_MISSING_TYPE | VIOLATION_KIND_UNDECLARED_PROPERTY)
-                ) {
-                Severity::Error
-            } else {
-                v.severity
-            };
-            let sev = match effective_severity {
-                Severity::Error => "error",
-                Severity::Warn => "warn",
-            };
-            // A missing/empty required property with no declared default cannot
-            // be synthesized by `--fix` (mapl BUG-3): tag it not-autofixable so
-            // the SCHEMA group reports `autofixable: false` unless some other
-            // SCHEMA violation in the file is fixable.
-            let autofixable = Some(v.kind != Some(VIOLATION_KIND_MISSING_REQUIRED_NO_DEFAULT));
-            violations_by_rule
-                .entry("SCHEMA".to_owned())
-                .or_default()
-                .push(InternalViolation {
-                    line: 1,
-                    column: 1,
-                    message: v.message,
-                    severity: sev.to_owned(),
-                    fix: None,
-                    fixed: false,
-                    autofixable,
-                });
-        }
-    }
-
-    // HYALO003 — date-format: frontmatter date-typed keys must hold YYYY-MM-DD values.
-    for diag in
-        engine.lint_frontmatter_hyalo003(rel_path, &properties, md_lint_config, rule_filter, strict)
-    {
-        let sev = match diag.severity {
-            hyalo_mdlint::DiagSeverity::Error => "error",
-            hyalo_mdlint::DiagSeverity::Warn => "warn",
-        };
-        violations_by_rule
-            .entry("HYALO003".to_owned())
-            .or_default()
-            .push(InternalViolation {
-                line: diag.line,
-                column: diag.column,
-                message: diag.message,
-                severity: sev.to_owned(),
-                fix: None,
-                fixed: false,
-                autofixable: None,
-            });
-    }
-
-    // HYALO004 — datetime-format: schema-declared datetime properties must
-    // hold `YYYY-MM-DDThh:mm:ss` values.
-    let doc_type_for_dt: Option<&str> = properties.get("type").and_then(|v| v.as_str());
-    let effective_schema_for_dt: TypeSchema = match doc_type_for_dt {
-        Some(t) => schema.merged_schema_for_type(t),
-        None => schema.default_schema().clone(),
-    };
-    let datetime_pairs: Vec<(&str, &str, bool)> = effective_schema_for_dt
-        .properties
-        .iter()
-        .filter_map(|(name, c)| {
-            let is_tz = match c {
-                PropertyConstraint::DateTime => false,
-                PropertyConstraint::DateTimeTz => true,
-                _ => return None,
-            };
-            let v = properties.get(name.as_str())?;
-            let s = v.as_str()?;
-            Some((name.as_str(), s, is_tz))
-        })
-        .collect();
-    for diag in engine.lint_frontmatter_hyalo004(
-        rel_path,
-        &datetime_pairs,
-        md_lint_config,
-        rule_filter,
-        strict,
-    ) {
-        let sev = match diag.severity {
-            hyalo_mdlint::DiagSeverity::Error => "error",
-            hyalo_mdlint::DiagSeverity::Warn => "warn",
-        };
-        violations_by_rule
-            .entry("HYALO004".to_owned())
-            .or_default()
-            .push(InternalViolation {
-                line: diag.line,
-                column: diag.column,
-                message: diag.message,
-                severity: sev.to_owned(),
-                fix: None,
-                fixed: false,
-                autofixable: None,
-            });
-    }
-
-    // Apply frontmatter fixes if requested.
-    let mut body_modified = false;
-    let mut fix_actions: Vec<FixAction> = Vec::new();
-    let mut post_fix_schema_remaining: Option<Vec<InternalViolation>> = None;
-    // Post-fix type (used for required_sections validation below). Defaults to the
-    // type in the unfixed frontmatter; apply_fixes may infer/insert a type via
-    // FRONTMATTER003, in which case we want to validate against that.
-    let mut post_fix_doc_type: Option<String> = properties
-        .get("type")
-        .and_then(|v| v.as_str())
-        .map(str::to_owned);
-    if matches!(fix, FixMode::Apply | FixMode::DryRun) {
-        let fix_all_rules = fix_rules.is_empty();
-        let should_fix_frontmatter = fix_all_rules
-            || fix_rules
-                .iter()
-                .any(|r| r == "SCHEMA" || r.starts_with("FRONTMATTER"));
-        if should_fix_frontmatter {
-            let mut mutable = properties.clone();
-            let actions = apply_fixes(rel_path, &mut mutable, schema);
-            if !actions.is_empty() {
-                let mut applied = true;
-                if matches!(fix, FixMode::Apply) {
-                    // A TOCTOU failure here (external edit during --fix) or
-                    // a write error is a real, expected runtime condition —
-                    // not exotic — and must not abort the whole batch via
-                    // `?` (M-1 follow-up, Finding 4, review round on PR
-                    // #254: only the initial read was hardened, not this
-                    // write path, so `lint_files_extended`'s merge loop
-                    // still killed the run on one file's write failure).
-                    // Catch it, report it as a diagnostic on this file, and
-                    // skip fixing (this file only) instead.
-                    let write_result = check_mtime(full_path, mtime0).and_then(|()| {
-                        write_frontmatter_within(vault_dir, full_path, &mutable)
-                            .with_context(|| format!("writing fixed frontmatter to {rel_path}"))
-                    });
-                    match write_result {
-                        Ok(()) => {
-                            // Re-baseline: the write above legitimately
-                            // changed the file's mtime, and a later
-                            // body-fix write in this same call must not
-                            // mistake it for a concurrent modification. If
-                            // *this* re-read fails, the frontmatter write
-                            // itself already succeeded — still count the
-                            // fix as applied, but skip attempting the body
-                            // fix below (no trustworthy baseline for its
-                            // own TOCTOU check).
-                            match read_mtime(full_path) {
-                                Ok(fresh) => mtime0 = fresh,
-                                Err(e) => {
-                                    frontmatter_write_failed = true;
-                                    push_fix_write_error_violation(
-                                        &mut violations_by_rule,
-                                        &format!(
-                                            "fixed frontmatter written to {rel_path}, but could \
-                                             not re-read its mtime for the body-fix TOCTOU \
-                                             check: {e}"
-                                        ),
-                                    );
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            applied = false;
-                            frontmatter_write_failed = true;
-                            push_fix_write_error_violation(
-                                &mut violations_by_rule,
-                                &format!("could not write fixed frontmatter to {rel_path}: {e}"),
-                            );
-                        }
-                    }
-                }
-                if applied {
-                    fix_actions = actions;
-                }
-            }
-            if should_include_frontmatter {
-                // Re-validate the mutated properties to get the actual
-                // post-fix SCHEMA remaining set. Avoids guessing via
-                // `fix_actions.len()`, which is not 1:1 with resolved
-                // diagnostics (one fix action can clear multiple violations,
-                // or insert defaults that don't clear any).
-                let post = validate_properties(rel_path, &mutable, schema, case_insensitive);
-                let remaining: Vec<InternalViolation> = post
-                    .into_iter()
-                    .map(|v| {
-                        let sev = match v.severity {
-                            Severity::Error => "error",
-                            Severity::Warn => "warn",
-                        };
-                        InternalViolation {
-                            line: 1,
-                            column: 1,
-                            message: v.message,
-                            severity: sev.to_owned(),
-                            fix: None,
-                            fixed: false,
-                            autofixable: None,
-                        }
-                    })
-                    .collect();
-                post_fix_schema_remaining = Some(remaining);
-            }
-            // Update post_fix_doc_type from the (possibly-inferred) mutable properties.
-            post_fix_doc_type = mutable
-                .get("type")
-                .and_then(|v| v.as_str())
-                .map(str::to_owned);
-        }
-    }
-
-    // Required-sections pass — only when rule_filter is empty or includes SCHEMA.
-    if should_include_frontmatter {
-        let effective_schema = match post_fix_doc_type.as_deref() {
-            Some(t) => schema.merged_schema_for_type(t),
-            None => schema.default_schema().clone(),
-        };
-        if !effective_schema.required_sections.is_empty() {
-            let section_violations = validate_required_sections(
-                full_path,
-                rel_path,
-                &effective_schema.required_sections,
-            )?;
-            for v in section_violations {
-                let sev = match v.severity {
-                    Severity::Error => "error",
-                    Severity::Warn => "warn",
-                };
-                violations_by_rule
-                    .entry("SCHEMA".to_owned())
-                    .or_default()
-                    .push(InternalViolation {
-                        line: 1,
-                        column: 1,
-                        message: v.message,
-                        severity: sev.to_owned(),
-                        fix: None,
-                        fixed: false,
-                        autofixable: None,
-                    });
-            }
-        }
-    }
-
-    // Body pass — extract frontmatter fields needed for HYALO rules.
-    let frontmatter_status = properties
-        .get("status")
-        .and_then(|v| v.as_str())
-        .map(str::to_owned);
-
-    // Track per-diagnostic outcomes for the new fix-mode JSON shape.
-    let mut body_fix_outcomes: Vec<(String, FixOutcome)> = Vec::new();
-    // Diagnostics that were fixed, accumulated across every pass below
-    // (each carries its own line/column, valid against the body revision it
-    // was found in — fine for display, since only `fixed`-derived counts
-    // and messages are surfaced, never the byte offsets).
-    let mut fixed_diagnostics: Vec<hyalo_mdlint::Diagnostic> = Vec::new();
-    // The body text, mutated in place across fix passes.
-    let mut working_body: String = body_content.to_owned();
-
-    // Re-lint and re-fix the working body up to `MAX_BODY_FIX_PASSES` times,
-    // or until nothing more changes — whichever comes first. Read-only mode
-    // and DryRun both run this same in-memory loop; only `FixMode::Apply`
-    // writes the result to disk (below), so DryRun still previews the
-    // fully-converged outcome.
-    let mut current_diagnostics = engine.lint_body(
-        &working_body,
-        rel_path,
-        frontmatter_status.as_deref(),
-        schema_has_completed,
-        md_lint_config,
-        rule_filter,
-    )?;
-
-    if matches!(fix, FixMode::Apply | FixMode::DryRun) {
-        let fix_all_rules = fix_rules.is_empty();
-
-        for _ in 0..MAX_BODY_FIX_PASSES {
-            if current_diagnostics.is_empty() {
-                break;
-            }
-            let fixable_indices: Vec<usize> = current_diagnostics
-                .iter()
-                .enumerate()
-                .filter(|(_, d)| d.fix.is_some())
-                .filter(|(_, d)| fix_all_rules || fix_rules.iter().any(|r| r == &d.rule_id))
-                .map(|(i, _)| i)
-                .collect();
-            if fixable_indices.is_empty() {
-                break;
-            }
-
-            let fixable_refs: Vec<&hyalo_mdlint::Diagnostic> = fixable_indices
-                .iter()
-                .map(|&i| &current_diagnostics[i])
-                .collect();
-            let (new_body, outcomes) = apply_body_fixes(&working_body, &fixable_refs);
-
-            let mut applied_this_pass: std::collections::HashSet<usize> =
-                std::collections::HashSet::new();
-            for (slot, &orig_idx) in fixable_indices.iter().enumerate() {
-                let rule_id = current_diagnostics[orig_idx].rule_id.clone();
-                match &outcomes[slot] {
-                    FixOutcome::Applied => {
-                        applied_this_pass.insert(orig_idx);
-                        body_fix_outcomes.push((rule_id, FixOutcome::Applied));
-                    }
-                    FixOutcome::Conflict { blocking_rule } => {
-                        body_fix_outcomes.push((
-                            rule_id,
-                            FixOutcome::Conflict {
-                                blocking_rule: blocking_rule.clone(),
-                            },
-                        ));
-                    }
-                    FixOutcome::NoFix => body_fix_outcomes.push((rule_id, FixOutcome::NoFix)),
-                }
-            }
-
-            if applied_this_pass.is_empty() {
-                // Every fixable diagnostic this pass hit a conflict or
-                // turned out to be a no-op — no progress possible, stop.
-                break;
-            }
-
-            for (i, d) in current_diagnostics.into_iter().enumerate() {
-                if applied_this_pass.contains(&i) {
-                    fixed_diagnostics.push(d);
-                }
-                // Diagnostics that weren't applied are dropped rather than
-                // carried forward: their byte offsets are stale against
-                // `new_body`, and any still-unresolved issue is rediscovered
-                // with correct positions by the re-lint below.
-            }
-
-            working_body = new_body;
-            current_diagnostics = engine.lint_body(
-                &working_body,
-                rel_path,
-                frontmatter_status.as_deref(),
-                schema_has_completed,
-                md_lint_config,
-                rule_filter,
-            )?;
-        }
-    }
-
-    // A TOCTOU failure (external edit during --fix) or an I/O error on any
-    // of the three fallible steps below (re-reading fresh frontmatter,
-    // `check_mtime`, or the write itself) is a real, expected runtime
-    // condition — not exotic — and must not abort the whole batch via `?`
-    // (M-1 follow-up, Finding 4, review round on PR #254: only the initial
-    // read was hardened, not this write path, so `lint_files_extended`'s
-    // merge loop still killed the run on one file's write failure here).
-    // Skipping the body write when the frontmatter write already failed
-    // above avoids compounding two partial-state failures on one file; the
-    // frontmatter failure's own diagnostic already covers it.
-    let mut body_write_failed = false;
-    if matches!(fix, FixMode::Apply) && working_body != body_content && !frontmatter_write_failed {
-        // Re-derive the frontmatter bytes fresh from disk when a
-        // frontmatter fix already landed above — `content[..body_start]` is
-        // a snapshot from before that write and would silently revert it if
-        // reused here.
-        let frontmatter_part: Result<Cow<'_, str>> = if fix_actions.is_empty() {
-            Ok(Cow::Borrowed(&content[..body_start]))
-        } else {
-            std::fs::read_to_string(full_path)
-                .with_context(|| format!("re-reading {rel_path} after frontmatter fix"))
-                .map(|fresh| {
-                    let fresh_body_start = find_body_start(&fresh);
-                    Cow::Owned(fresh[..fresh_body_start].to_owned())
-                })
-        };
-        let write_result = frontmatter_part.and_then(|frontmatter_part| {
-            check_mtime(full_path, mtime0)?;
-            let new_content = format!("{frontmatter_part}{working_body}");
-            hyalo_core::atomic_write_within(vault_dir, full_path, new_content.as_bytes())
-                .with_context(|| format!("writing fixed body to {rel_path}"))
-        });
-        match write_result {
-            Ok(()) => body_modified = true,
-            Err(e) => {
-                body_write_failed = true;
-                push_fix_write_error_violation(
-                    &mut violations_by_rule,
-                    &format!("could not write fixed body to {rel_path}: {e}"),
-                );
-            }
-        }
-    }
-
-    // Body rules lint the post-frontmatter slice, so their diagnostics carry
-    // body-relative line numbers. Translate them to file-absolute lines so a
-    // reported `line N` matches the raw file (BUG-6). `body_line_offset` is the
-    // 1-based file line on which the body begins; body-relative line `L` maps
-    // to `L + offset - 1`. With no frontmatter `offset == 1`, i.e. a no-op.
-    let body_line_offset = find_body_line_offset(&content, body_start);
-    let to_file_line =
-        |body_line: usize| body_line.saturating_add(body_line_offset.saturating_sub(1));
-
-    // Group body diagnostics by rule: violations fixed across any pass,
-    // followed by whatever remains after the loop above (or the single
-    // read-only lint pass, when fix-mode is off).
-    let diag_to_violation = |d: hyalo_mdlint::Diagnostic, fixed: bool| {
-        let fix = d.fix.as_ref().map(|f| {
-            serde_json::json!({
-                "description": f.description,
-                "start": f.start,
-                "end": f.end,
-                "replacement": f.replacement,
-            })
-        });
-        InternalViolation {
-            line: to_file_line(d.line),
-            column: d.column,
-            message: d.message,
-            severity: format!("{}", d.severity),
-            fix,
-            fixed,
-            autofixable: None,
-        }
-    };
-    for d in fixed_diagnostics {
-        let rule_id = d.rule_id.clone();
-        // `fixed_diagnostics` was populated by the in-memory fix loop above,
-        // before it was known whether the write to disk would succeed. If
-        // it didn't (`body_write_failed`), nothing was actually fixed — the
-        // report must not claim `fixed: true` for a change that only ever
-        // existed in memory (Finding 4, review round on PR #254).
-        violations_by_rule
-            .entry(rule_id)
-            .or_default()
-            .push(diag_to_violation(d, !body_write_failed));
-    }
-    for d in current_diagnostics {
-        let rule_id = d.rule_id.clone();
-        violations_by_rule
-            .entry(rule_id)
-            .or_default()
-            .push(diag_to_violation(d, false));
-    }
-
-    // HYALO006 (broken-link) — vault-aware rule. Runs only when a shared
-    // LinkLintContext was built for this invocation (it is skipped when the
-    // rule is disabled or filtered out, so no context is constructed). Each of
-    // the file's links is resolved through the single shared resolver entry
-    // point; unresolved ones become findings. Severity follows the
-    // configurable / --strict pattern used by the other HYALO rules.
-    if let Some(link_ctx) = link_ctx {
-        let base_sev = md_lint_config
-            .rules
-            .get(RULE_ID_BROKEN_LINK)
-            .and_then(hyalo_mdlint::RuleOverride::severity)
-            .map_or("warn", |s| {
-                if s.eq_ignore_ascii_case("error") {
-                    "error"
-                } else {
-                    "warn"
-                }
-            });
-        // --strict promotes the default warn to error (unless the user pinned
-        // an explicit severity via `[lint.rules.HYALO006] severity`).
-        let has_explicit_sev = md_lint_config
-            .rules
-            .get(RULE_ID_BROKEN_LINK)
-            .and_then(hyalo_mdlint::RuleOverride::severity)
-            .is_some();
-        let severity = if strict && !has_explicit_sev {
-            "error"
-        } else {
-            base_sev
-        };
-        for f in
-            hyalo_mdlint::profiles::link::check_broken_links(link_ctx, content.as_bytes(), rel_path)
-        {
-            violations_by_rule
-                .entry(RULE_ID_BROKEN_LINK.to_owned())
-                .or_default()
-                .push(InternalViolation {
-                    // iter-211 / BUG-9: HYALO006 scans the WHOLE file (frontmatter
-                    // included) through `scan_slice_multi`, whose visitor line
-                    // numbers are already file-absolute. The body-rule
-                    // `to_file_line` translation must NOT be applied here — doing
-                    // so added the frontmatter length a second time, so a link on
-                    // line 5 of a file with 3 frontmatter lines was reported at
-                    // line 8. MD rules keep `to_file_line` because they lint the
-                    // post-frontmatter slice.
-                    line: f.line,
-                    column: 1,
-                    message: f.message,
-                    severity: severity.to_owned(),
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                });
-        }
-    }
-
-    // OKF conformance profile — advisory (warn-level) rules layered on top of
-    // the schema pass. Only runs under `hyalo lint --profile okf`. Every OKF
-    // rule respects `[lint.rules]` enable/disable overrides and the
-    // `--rule`/`--rule-prefix` filters, exactly like the HYALO rules above.
-    if okf_profile {
-        let is_enabled = |rule_id: &str| -> bool {
-            // `[lint.rules.<id>] enabled = false` disables it; default on.
-            let cfg_enabled = md_lint_config
-                .rules
-                .get(rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::enabled)
-                .unwrap_or(true);
-            if !cfg_enabled {
-                return false;
-            }
-            // Honor --rule / --rule-prefix (rule_filter is the resolved id set).
-            rule_filter.is_empty() || rule_filter.iter().any(|r| r == rule_id)
-        };
-        let okf_doc_type = properties.get("type").and_then(|v| v.as_str());
-        let findings = hyalo_mdlint::profiles::okf::run_okf_rules(
-            rel_path,
-            full_path,
-            &content,
-            body_content,
-            find_body_line_offset(&content, body_start),
-            okf_doc_type,
-            &is_enabled,
-            vault_dir,
-            case_insensitive,
-        );
-        for f in findings {
-            // Apply per-rule severity override; OKF rules default to warn.
-            let severity = md_lint_config
-                .rules
-                .get(f.rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::severity)
-                .unwrap_or("warn")
-                .to_owned();
-            violations_by_rule
-                .entry(f.rule_id.to_owned())
-                .or_default()
-                .push(InternalViolation {
-                    line: f.line,
-                    column: 1,
-                    message: f.message,
-                    severity,
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                });
-        }
-    }
-
-    // MADR conformance profile — advisory (warn-level) rules layered on top of
-    // the schema pass. Only runs under `hyalo lint --profile madr` (or a vault
-    // whose `.hyalo.toml` sets `[lint] profile = "madr"`). Same override/filter
-    // discipline as the OKF block above.
-    if madr_profile {
-        let is_enabled = |rule_id: &str| -> bool {
-            let cfg_enabled = md_lint_config
-                .rules
-                .get(rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::enabled)
-                .unwrap_or(true);
-            if !cfg_enabled {
-                return false;
-            }
-            rule_filter.is_empty() || rule_filter.iter().any(|r| r == rule_id)
-        };
-        // Effective type: explicit `type:` frontmatter else the path binding.
-        let explicit_type = properties.get("type").and_then(|v| v.as_str());
-        let effective_type = explicit_type.or_else(|| schema.bound_type_for(rel_path));
-        let status = properties.get("status").and_then(|v| v.as_str());
-        let findings = hyalo_mdlint::profiles::madr::run_madr_rules(
-            rel_path,
-            full_path,
-            effective_type,
-            status,
-            &is_enabled,
-        );
-        for f in findings {
-            let severity = md_lint_config
-                .rules
-                .get(f.rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::severity)
-                .unwrap_or("warn")
-                .to_owned();
-            violations_by_rule
-                .entry(f.rule_id.to_owned())
-                .or_default()
-                .push(InternalViolation {
-                    line: f.line,
-                    column: 1,
-                    message: f.message,
-                    severity,
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                });
-        }
-    }
-
-    // Agent Skills conformance profile — rules layered on top of the schema
-    // pass. Only runs under `hyalo lint --profile skills` (or a vault whose
-    // `.hyalo.toml` sets `[lint] profile = "skills"`). Mostly advisory
-    // (warn-level): `SKILL-RESERVED-NAME` is the exception and defaults to
-    // error severity (a reserved `name` is a hard spec violation). Same
-    // override/filter discipline as the OKF/MADR blocks above.
-    if skills_profile {
-        let is_enabled = |rule_id: &str| -> bool {
-            let cfg_enabled = md_lint_config
-                .rules
-                .get(rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::enabled)
-                .unwrap_or(true);
-            if !cfg_enabled {
-                return false;
-            }
-            rule_filter.is_empty() || rule_filter.iter().any(|r| r == rule_id)
-        };
-        // Effective type: explicit `type:` frontmatter else the path binding.
-        let explicit_type = properties.get("type").and_then(|v| v.as_str());
-        let effective_type = explicit_type.or_else(|| schema.bound_type_for(rel_path));
-        let name = properties.get("name").and_then(|v| v.as_str());
-        // Line budget counts the markdown body only (frontmatter excluded). A
-        // trailing newline does not add a phantom final line.
-        let body_line_count = body_content.lines().count();
-        let findings = hyalo_mdlint::profiles::skills::run_skill_rules(
-            rel_path,
-            effective_type,
-            name,
-            body_line_count,
-            &is_enabled,
-        );
-        for f in findings {
-            let severity = md_lint_config
-                .rules
-                .get(f.rule_id)
-                .and_then(hyalo_mdlint::RuleOverride::severity)
-                .unwrap_or(f.default_severity)
-                .to_owned();
-            violations_by_rule
-                .entry(f.rule_id.to_owned())
-                .or_default()
-                .push(InternalViolation {
-                    line: f.line,
-                    column: 1,
-                    message: f.message,
-                    severity,
-                    fix: None,
-                    fixed: false,
-                    autofixable: None,
-                });
-        }
-    }
-
-    // Keep a Changelog conformance profile — the grammar rules layered on top
-    // of the schema pass. Only runs under `hyalo lint --profile changelog` (or a
-    // vault whose `.hyalo.toml` sets `[lint] profile = "changelog"`). Rules
-    // dispatch only on the file whose effective type is `changelog` (the literal
-    // `CHANGELOG.md` binding). Severity is *mixed*: the grammar/ordering rules
-    // default to error, the empty-section and link-ref cross-check to warn — the
-    // per-finding `default_severity` is the source of truth, honoured below.
-    // Same override/filter discipline as the OKF/MADR/skills blocks above.
-    if changelog_profile {
-        let explicit_type = properties.get("type").and_then(|v| v.as_str());
-        let effective_type = explicit_type.or_else(|| schema.bound_type_for(rel_path));
-        if matches!(effective_type, Some(t) if t.eq_ignore_ascii_case("changelog")) {
-            let is_enabled = |rule_id: &str| -> bool {
-                let cfg_enabled = md_lint_config
-                    .rules
-                    .get(rule_id)
-                    .and_then(hyalo_mdlint::RuleOverride::enabled)
-                    .unwrap_or(true);
-                if !cfg_enabled {
-                    return false;
-                }
-                rule_filter.is_empty() || rule_filter.iter().any(|r| r == rule_id)
-            };
-            let findings = hyalo_mdlint::profiles::changelog::run_changelog_rules(
-                &content,
-                body_content,
-                find_body_line_offset(&content, body_start),
-                &is_enabled,
-            );
-            for f in findings {
-                let severity = md_lint_config
-                    .rules
-                    .get(f.rule_id)
-                    .and_then(hyalo_mdlint::RuleOverride::severity)
-                    .unwrap_or(f.default_severity)
-                    .to_owned();
-                violations_by_rule
-                    .entry(f.rule_id.to_owned())
-                    .or_default()
-                    .push(InternalViolation {
-                        line: f.line,
-                        column: 1,
-                        message: f.message,
-                        severity,
-                        fix: None,
-                        fixed: false,
-                        autofixable: None,
-                    });
-            }
-        }
-    }
-
-    // Mirror the `fixed_diagnostics` correction above: `body_fix_outcomes`
-    // was recorded by the in-memory fix loop before the write's outcome was
-    // known, so an `Applied` entry here must be downgraded to `NoFix` when
-    // the write to disk never actually happened (Finding 4, review round on
-    // PR #254) — otherwise fix-mode's totals (`fix_mode_file_totals`) would
-    // count a change that was never persisted.
-    if body_write_failed {
-        for (_, outcome) in &mut body_fix_outcomes {
-            if matches!(outcome, FixOutcome::Applied) {
-                *outcome = FixOutcome::NoFix;
-            }
-        }
-    }
-
-    let total_violations = violations_by_rule.values().map(Vec::len).sum();
-
-    let _ = max_per_rule; // applied during output construction
-
-    Ok(PerFileLintResult {
-        rel_path: rel_path.to_owned(),
-        doc_type: post_fix_doc_type,
-        violations_by_rule,
-        total_violations,
-        body_modified,
-        fix_actions,
-        body_fix_outcomes,
-        post_fix_schema_remaining,
-    })
-}
-
-/// Apply body fixes greedily.
-///
-/// Returns `(fixed_content, outcomes)` where `outcomes[i]` corresponds to
-/// `fixes[i]` — either `Applied`, `Conflict`, or (for diagnostics without a
-/// fix, a fix that lost a conflict, or a fix that turned out to be a
-/// byte-for-byte no-op) `NoFix`.
-///
-/// Conflict resolution and buffer mutation use two different orderings on
-/// purpose:
-/// - **Winner selection** happens in priority order (`Error` before `Warn`,
-///   then descending start offset), so a higher-severity fix (e.g. HYALO001)
-///   is never displaced by an overlapping lower-severity one (e.g. MD009)
-///   just because the latter happens to sort first by offset.
-/// - **Buffer mutation** of the resulting non-overlapping winners always
-///   proceeds in descending start-offset order, which is required for
-///   correctness: each edit's range must still be valid against the
-///   partially mutated buffer, and that only holds if edits at higher
-///   offsets (later in the string) are applied first.
-fn apply_body_fixes(body: &str, fixes: &[&hyalo_mdlint::Diagnostic]) -> (String, Vec<FixOutcome>) {
-    let severity_rank = |sev: hyalo_mdlint::DiagSeverity| match sev {
-        hyalo_mdlint::DiagSeverity::Error => 0,
-        hyalo_mdlint::DiagSeverity::Warn => 1,
-    };
-
-    // (original_index, start, end), ordered by selection priority.
-    let mut candidates: Vec<(usize, usize, usize)> = fixes
-        .iter()
-        .enumerate()
-        .filter_map(|(i, d)| d.fix.as_ref().map(|f| (i, f.start, f.end)))
-        .collect();
-    candidates.sort_by(|&(ia, sa, _), &(ib, sb, _)| {
-        severity_rank(fixes[ia].severity)
-            .cmp(&severity_rank(fixes[ib].severity))
-            .then(sb.cmp(&sa))
-    });
-
-    let mut winners: Vec<(usize, usize, usize)> = Vec::new(); // (orig_idx, start, end)
-    let mut outcome_map: std::collections::HashMap<usize, FixOutcome> =
-        std::collections::HashMap::new();
-
-    for &(orig_idx, start, end) in &candidates {
-        let conflict_with = winners.iter().find(|(_, ws, we)| start < *we && end > *ws);
-        if let Some(&(blocking_idx, _, _)) = conflict_with {
-            let blocking_rule = fixes[blocking_idx].rule_id.clone();
-            outcome_map.insert(orig_idx, FixOutcome::Conflict { blocking_rule });
-            continue;
-        }
-        // `start > end`, a non-char-boundary offset, or `end` past the body all
-        // make the range unusable — `str` indexing would panic on any of them.
-        // `get` is the single check that covers all three (iter-191).
-        if body.get(start..end).is_none() {
-            outcome_map.insert(
-                orig_idx,
-                FixOutcome::Conflict {
-                    blocking_rule: "out-of-bounds".to_owned(),
-                },
-            );
-            continue;
-        }
-        winners.push((orig_idx, start, end));
-    }
-
-    // Mutate the buffer in descending start order (see doc comment above).
-    winners.sort_by_key(|&(_, start, _)| std::cmp::Reverse(start));
-    let mut result = body.to_owned();
-    for (orig_idx, start, end) in winners {
-        let replacement = fixes[orig_idx]
-            .fix
-            .as_ref()
-            .map_or("", |f| f.replacement.as_str());
-        // Re-validate against the *partially mutated* buffer: the winner
-        // selection above checked the range against the pristine body, and
-        // although descending-offset mutation keeps earlier ranges valid, a
-        // `get` here is what makes that reasoning enforced rather than
-        // assumed. `None` means the range is unusable — report it as a
-        // conflict instead of panicking on a slice (iter-191).
-        let Some(current) = result.get(start..end) else {
-            outcome_map.insert(
-                orig_idx,
-                FixOutcome::Conflict {
-                    blocking_rule: "out-of-bounds".to_owned(),
-                },
-            );
-            continue;
-        };
-        if current == replacement {
-            // Byte-for-byte no-op: nothing changed, don't count it as fixed.
-            outcome_map.insert(orig_idx, FixOutcome::NoFix);
-            continue;
-        }
-        result.replace_range(start..end, replacement);
-        outcome_map.insert(orig_idx, FixOutcome::Applied);
-    }
-
-    let outcomes: Vec<FixOutcome> = (0..fixes.len())
-        .map(|i| outcome_map.remove(&i).unwrap_or(FixOutcome::NoFix))
-        .collect();
-
-    (result, outcomes)
-}
-
-/// Find the byte offset where the document body starts (after the closing `---` line).
-/// Returns 0 if no frontmatter is found.
-///
-/// The opening check mirrors the shared frontmatter policy (iter-158 C-1):
-/// an optional single UTF-8 BOM, then a line that is exactly `---` — so lint
-/// splits BOM-prefixed files the same way the read/write paths do instead of
-/// treating the whole file as body.
-fn find_body_start(content: &str) -> usize {
-    let rest = content.strip_prefix('\u{feff}').unwrap_or(content);
-    if !(rest.starts_with("---\n") || rest.starts_with("---\r\n") || rest == "---") {
-        return 0;
-    }
-    // Find the second `---` delimiter.
-    let after_first = content.find('\n').map_or(content.len(), |i| i + 1);
-    let rest = &content[after_first..];
-    if let Some(pos) = rest.find("\n---") {
-        // Skip past `\n---\n` or `\n---` at end.
-        let abs = after_first + pos + 4; // skip \n---
-        // Skip the terminator after the closing `---` — LF or CRLF, so the
-        // body slice never starts with a stray carriage return on CRLF files.
-        let bytes = content.as_bytes();
-        if bytes.get(abs) == Some(&b'\r') && bytes.get(abs + 1) == Some(&b'\n') {
-            abs + 2
-        } else if bytes.get(abs) == Some(&b'\n') {
-            abs + 1
-        } else {
-            abs
-        }
-    } else {
-        // No closing delimiter — treat whole file as body.
-        0
-    }
-}
-
-/// Compute the 1-based file line number on which the body begins, given the
-/// byte offset returned by [`find_body_start`]. Used so OKF findings report
-/// file-absolute line numbers rather than body-relative ones.
-fn find_body_line_offset(content: &str, body_start: usize) -> usize {
-    // Line 1 when there is no frontmatter; otherwise 1 + number of newlines
-    // consumed by the frontmatter block.
-    1 + content[..body_start]
-        .bytes()
-        .filter(|b| *b == b'\n')
-        .count()
-}
-
-/// Check whether any schema type declares `status` as an enum with `completed`.
-fn schema_has_completed_status(schema: &SchemaConfig) -> bool {
-    // Check default schema.
-    if has_completed_in_type(&schema.default_schema().properties) {
-        return true;
-    }
-    // Check all typed schemas.
-    for ts in schema.types.values() {
-        if has_completed_in_type(&ts.properties) {
-            return true;
-        }
-    }
-    false
-}
-
-fn has_completed_in_type(props: &std::collections::HashMap<String, PropertyConstraint>) -> bool {
-    if let Some(PropertyConstraint::Enum { values }) = props.get("status") {
-        return values.iter().any(|v| v == "completed");
-    }
-    false
-}
-
-// ---------------------------------------------------------------------------
-// Unit tests
-// ---------------------------------------------------------------------------
+//
+// `commands/lint.rs` had grown to 4,005 lines -- a review hotspot flagged by
+// the 2026-08-27 deep review. It is split by pass: the config-level checks, the
+// output shapes, the vault-level pass, the per-file pass, the body-fix
+// application, and the dispatch arm. This file keeps only the shared imports
+// and re-exports, so `commands::lint::` call sites are unaffected.
+
+mod config_checks;
+mod engine;
+mod file;
+mod fix;
+mod run;
+mod types;
+
+// Glob re-exports rather than named lists: every item below kept the exact
+// visibility it had inside the pre-split module, and a glob re-export cannot
+// widen it (a `pub(super)` item stays `pub(super)`). That is what makes the
+// split invisible to callers -- and it is also how the submodules see each
+// other, since each one imports this parent.
+pub use config_checks::*;
+pub use engine::*;
+use file::lint_one_file_extended;
+use fix::{apply_body_fixes, find_body_line_offset, find_body_start, schema_has_completed_status};
+pub use types::*;
+
+pub(crate) use run::run;
 
 #[cfg(test)]
-#[allow(clippy::items_after_test_module)] // dispatch handler appended below (ARCH-1, iter-225)
-mod tests {
-    use super::*;
-    use hyalo_core::is_iso8601_date;
-    use hyalo_core::schema::{PropertyConstraint, SchemaConfig, TypeSchema};
-    use indexmap::IndexMap;
-    use serde_json::Value;
-    use std::collections::HashMap;
-
-    fn make_schema(
-        default_required: &[&str],
-        type_name: &str,
-        type_required: &[&str],
-        type_properties: HashMap<&str, PropertyConstraint>,
-    ) -> SchemaConfig {
-        let default = TypeSchema {
-            required: default_required.iter().map(ToString::to_string).collect(),
-            ..Default::default()
-        };
-        let mut props: HashMap<String, PropertyConstraint> = HashMap::new();
-        for (k, v) in type_properties {
-            props.insert(k.to_owned(), v);
-        }
-        let type_schema = TypeSchema {
-            required: type_required.iter().map(ToString::to_string).collect(),
-            properties: props,
-            ..Default::default()
-        };
-        let mut types = HashMap::new();
-        types.insert(type_name.to_owned(), type_schema);
-        SchemaConfig {
-            default,
-            types,
-            ..Default::default()
-        }
-    }
-
-    // --- validate_schema_config (review round finding 2) ---
-
-    #[test]
-    fn validate_schema_config_none_when_valid() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join(".hyalo.toml"),
-            "dir = \".\"\n[schema.types.task.properties.priority]\ntype = \"number\"\nminimum = 1\nmaximum = 5\n",
-        )
-        .unwrap();
-        assert!(validate_schema_config(dir.path(), false).is_none());
-        assert!(validate_schema_config(dir.path(), true).is_none());
-    }
-
-    #[test]
-    fn validate_schema_config_none_when_no_schema_block() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join(".hyalo.toml"), "dir = \".\"\n").unwrap();
-        assert!(validate_schema_config(dir.path(), false).is_none());
-    }
-
-    #[test]
-    fn validate_schema_config_none_when_no_config_file() {
-        let dir = tempfile::tempdir().unwrap();
-        assert!(validate_schema_config(dir.path(), false).is_none());
-    }
-
-    #[test]
-    fn validate_schema_config_warn_severity_without_strict() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join(".hyalo.toml"),
-            "dir = \".\"\n[schema.types.task.properties.title]\ntype = \"string\"\npatterns = \".*\"\n",
-        )
-        .unwrap();
-        let result = validate_schema_config(dir.path(), false).expect("malformed schema");
-        assert_eq!(result.file, ".hyalo.toml");
-        assert_eq!(result.violations.len(), 1);
-        assert_eq!(result.violations[0].severity, Severity::Warn);
-        assert_eq!(
-            result.violations[0].kind,
-            Some(VIOLATION_KIND_SCHEMA_MALFORMED)
-        );
-        assert!(result.violations[0].message.contains("patterns"));
-    }
-
-    #[test]
-    fn validate_schema_config_error_severity_under_strict() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join(".hyalo.toml"),
-            "dir = \".\"\n[schema.types.task.properties.title]\ntype = \"string\"\npatterns = \".*\"\n",
-        )
-        .unwrap();
-        let result = validate_schema_config(dir.path(), true).expect("malformed schema");
-        assert_eq!(result.violations[0].severity, Severity::Error);
-    }
-
-    // --- is_iso8601_date ---
-
-    #[test]
-    fn valid_date() {
-        assert!(is_iso8601_date("2026-04-13"));
-    }
-
-    #[test]
-    fn normalize_date_padding_and_calendar() {
-        // Short month/day get zero-padded.
-        assert_eq!(normalize_date("2026-4-9"), Some("2026-04-09".to_owned()));
-        // Feb 29 is valid in leap years only.
-        assert_eq!(normalize_date("2024-2-29"), Some("2024-02-29".to_owned()));
-        assert_eq!(normalize_date("2023-2-29"), None);
-        // Out-of-range days/months are rejected, not silently normalized.
-        assert_eq!(normalize_date("2026-02-31"), None);
-        assert_eq!(normalize_date("2026-04-31"), None);
-        assert_eq!(normalize_date("2026-13-01"), None);
-    }
-
-    #[test]
-    fn invalid_date_format() {
-        assert!(!is_iso8601_date("April 13"));
-        assert!(!is_iso8601_date("13-04-2026"));
-        assert!(!is_iso8601_date("2026/04/13"));
-    }
-
-    // Test helper: wraps `validate_constraint` with a throwaway regex cache.
-    // Returns the first violation (or None) for constraints that produce at most one.
-    fn vc(name: &str, value: &Value, c: &PropertyConstraint) -> Option<Violation> {
-        let mut cache = HashMap::new();
-        validate_constraint(name, value, c, &mut cache)
-            .into_iter()
-            .next()
-    }
-
-    // Test helper: returns all violations from `validate_constraint`.
-    fn vc_all(name: &str, value: &Value, c: &PropertyConstraint) -> Vec<Violation> {
-        let mut cache = HashMap::new();
-        validate_constraint(name, value, c, &mut cache)
-    }
-
-    // --- validate_constraint ---
-
-    #[test]
-    fn date_constraint_valid() {
-        let v = vc(
-            "date",
-            &Value::String("2026-04-13".into()),
-            &PropertyConstraint::Date,
-        );
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn date_constraint_invalid() {
-        let v = vc(
-            "date",
-            &Value::String("April 13".into()),
-            &PropertyConstraint::Date,
-        );
-        assert!(matches!(
-            v,
-            Some(Violation {
-                severity: Severity::Error,
-                ..
-            })
-        ));
-    }
-
-    #[test]
-    fn enum_constraint_valid() {
-        let v = vc(
-            "status",
-            &Value::String("planned".into()),
-            &PropertyConstraint::Enum {
-                values: vec!["planned".into(), "done".into()],
-            },
-        );
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn enum_constraint_invalid_with_suggestion() {
-        let v = vc(
-            "status",
-            &Value::String("planed".into()),
-            &PropertyConstraint::Enum {
-                values: vec!["planned".into(), "done".into()],
-            },
-        );
-        let viol = v.expect("expected violation");
-        assert_eq!(viol.severity, Severity::Error);
-        assert!(viol.message.contains("did you mean \"planned\""));
-    }
-
-    #[test]
-    fn number_constraint_valid() {
-        let v = vc(
-            "priority",
-            &Value::Number(5.into()),
-            &PropertyConstraint::Number {
-                minimum: None,
-                maximum: None,
-            },
-        );
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn number_constraint_invalid() {
-        let v = vc(
-            "priority",
-            &Value::String("five".into()),
-            &PropertyConstraint::Number {
-                minimum: None,
-                maximum: None,
-            },
-        );
-        assert!(matches!(
-            v,
-            Some(Violation {
-                severity: Severity::Error,
-                ..
-            })
-        ));
-    }
-
-    #[test]
-    fn number_constraint_minimum_violation() {
-        let v = vc(
-            "priority",
-            &Value::Number(0.into()),
-            &PropertyConstraint::Number {
-                minimum: Some(1.0),
-                maximum: None,
-            },
-        );
-        let viol = v.expect("expected a minimum violation");
-        assert_eq!(viol.severity, Severity::Error);
-        assert!(viol.message.contains("minimum"));
-    }
-
-    #[test]
-    fn number_constraint_maximum_violation() {
-        let v = vc(
-            "priority",
-            &Value::Number(99.into()),
-            &PropertyConstraint::Number {
-                minimum: None,
-                maximum: Some(5.0),
-            },
-        );
-        let viol = v.expect("expected a maximum violation");
-        assert_eq!(viol.severity, Severity::Error);
-        assert!(viol.message.contains("maximum"));
-    }
-
-    #[test]
-    fn number_constraint_within_min_max_range_is_valid() {
-        let v = vc(
-            "priority",
-            &Value::Number(3.into()),
-            &PropertyConstraint::Number {
-                minimum: Some(1.0),
-                maximum: Some(5.0),
-            },
-        );
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn number_constraint_at_min_max_boundary_is_valid() {
-        // Bounds are inclusive.
-        assert!(
-            vc(
-                "priority",
-                &Value::Number(1.into()),
-                &PropertyConstraint::Number {
-                    minimum: Some(1.0),
-                    maximum: Some(5.0),
-                },
-            )
-            .is_none()
-        );
-        assert!(
-            vc(
-                "priority",
-                &Value::Number(5.into()),
-                &PropertyConstraint::Number {
-                    minimum: Some(1.0),
-                    maximum: Some(5.0),
-                },
-            )
-            .is_none()
-        );
-    }
-
-    #[test]
-    fn boolean_constraint_valid() {
-        let v = vc("draft", &Value::Bool(true), &PropertyConstraint::Boolean);
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn boolean_constraint_invalid() {
-        let v = vc(
-            "draft",
-            &Value::String("yes".into()),
-            &PropertyConstraint::Boolean,
-        );
-        assert!(matches!(
-            v,
-            Some(Violation {
-                severity: Severity::Error,
-                ..
-            })
-        ));
-    }
-
-    #[test]
-    fn list_constraint_valid() {
-        let v = vc("tags", &Value::Array(vec![]), &PropertyConstraint::List);
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn list_constraint_invalid() {
-        let v = vc(
-            "tags",
-            &Value::String("rust".into()),
-            &PropertyConstraint::List,
-        );
-        assert!(matches!(
-            v,
-            Some(Violation {
-                severity: Severity::Error,
-                ..
-            })
-        ));
-    }
-
-    #[test]
-    fn string_pattern_constraint_valid() {
-        let v = vc(
-            "branch",
-            &Value::String("iter-42/my-feature".into()),
-            &PropertyConstraint::String {
-                pattern: Some(r"^iter-\d+/".into()),
-                min_length: None,
-                max_length: None,
-            },
-        );
-        assert!(v.is_none());
-    }
-
-    #[test]
-    fn string_pattern_constraint_invalid() {
-        let v = vc(
-            "branch",
-            &Value::String("feature/my-branch".into()),
-            &PropertyConstraint::String {
-                pattern: Some(r"^iter-\d+/".into()),
-                min_length: None,
-                max_length: None,
-            },
-        );
-        assert!(matches!(
-            v,
-            Some(Violation {
-                severity: Severity::Error,
-                ..
-            })
-        ));
-    }
-
-    // --- lint_file via a temp file ---
-
-    #[test]
-    fn lint_file_missing_required() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(&path, "---\ntitle: Hello\n---\nBody\n").unwrap();
-
-        let schema = make_schema(&["title", "date"], "note", &[], HashMap::new());
-        let result = lint_file(&path, "note.md", &schema, false).unwrap();
-        // date is in default required, but only "title" is present.
-        // No type -> warn about no type. date missing -> error.
-        assert!(
-            result
-                .violations
-                .iter()
-                .any(|v| v.severity == Severity::Error
-                    && v.message.contains("missing required property \"date\""))
-        );
-    }
-
-    #[test]
-    fn missing_required_no_default_is_not_autofixable() {
-        // mapl BUG-3: a missing required property with no schema default cannot
-        // be synthesized by --fix, so its violation is tagged not-autofixable.
-        let schema = make_schema(&["title", "date"], "note", &[], HashMap::new());
-        let props: IndexMap<String, Value> = IndexMap::new(); // nothing present
-        let violations = validate_properties("note.md", &props, &schema, false);
-        let date_v = violations
-            .iter()
-            .find(|v| v.message.contains("missing required property \"date\""))
-            .expect("date must be flagged");
-        assert_eq!(
-            date_v.kind,
-            Some(VIOLATION_KIND_MISSING_REQUIRED_NO_DEFAULT),
-            "no-default missing-required is tagged not-autofixable"
-        );
-    }
-
-    #[test]
-    fn missing_required_with_default_is_autofixable() {
-        // When the schema declares a default, --fix CAN synthesize the value, so
-        // the violation is NOT tagged with the no-default kind.
-        let mut default = TypeSchema {
-            required: vec!["title".to_owned(), "status".to_owned()],
-            ..Default::default()
-        };
-        default
-            .defaults
-            .insert("status".to_owned(), "draft".to_owned());
-        let schema = SchemaConfig {
-            default,
-            ..Default::default()
-        };
-        let props: IndexMap<String, Value> = IndexMap::new();
-        let violations = validate_properties("note.md", &props, &schema, false);
-        let status_v = violations
-            .iter()
-            .find(|v| v.message.contains("missing required property \"status\""))
-            .expect("status must be flagged");
-        assert_ne!(
-            status_v.kind,
-            Some(VIOLATION_KIND_MISSING_REQUIRED_NO_DEFAULT),
-            "a defaulted missing-required stays autofixable"
-        );
-    }
-
-    #[test]
-    fn lint_file_no_type_warn() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(&path, "---\ntitle: Hello\n---\nBody\n").unwrap();
-
-        let schema = make_schema(&["title"], "note", &[], HashMap::new());
-        let result = lint_file(&path, "note.md", &schema, false).unwrap();
-        assert!(
-            result
-                .violations
-                .iter()
-                .any(|v| v.severity == Severity::Warn && v.message.contains("no 'type' property"))
-        );
-    }
-
-    /// Build a `SchemaConfig` from a TOML `[schema]` fragment (as it would
-    /// appear in `.hyalo.toml`), for tests that exercise bind = typing.
-    fn schema_from_toml(fragment: &str) -> SchemaConfig {
-        let val: toml::Value = toml::from_str(fragment).expect("valid schema fragment");
-        let raw: hyalo_core::schema::RawSchemaConfig = val
-            .get("schema")
-            .and_then(|v| v.clone().try_into().ok())
-            .expect("schema section present");
-        SchemaConfig::try_from(raw).expect("valid schema")
-    }
-
-    #[test]
-    fn bind_typed_frontmatterless_file_satisfies_required_type() {
-        // iter-172 bind = typing: a file whose type comes from a `[schema.bind]`
-        // path binding must satisfy `[schema.default] required = ["type"]`
-        // WITHOUT explicit `type:` frontmatter — a spec-valid frontmatter-less
-        // SKILL.md / ADR must lint clean under composed profiles.
-        let schema = schema_from_toml(
-            "\
-[schema.default]
-required = [\"type\"]
-
-[schema.types.skill]
-required = [\"name\", \"description\"]
-
-[[schema.bind]]
-glob = \"**/SKILL.md\"
-type = \"skill\"
-",
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("SKILL.md");
-        // A skill with the two required fields but NO explicit `type:`.
-        std::fs::write(
-            &path,
-            "---\nname: my-skill\ndescription: does a thing\n---\nBody\n",
-        )
-        .unwrap();
-
-        let result = lint_file(&path, "foo/SKILL.md", &schema, false).unwrap();
-        assert!(
-            !result
-                .violations
-                .iter()
-                .any(|v| v.message.contains("missing required property \"type\"")),
-            "bound file must not require explicit type: {:?}",
-            result.violations
-        );
-        assert!(
-            !result
-                .violations
-                .iter()
-                .any(|v| v.message.contains("no 'type' property")),
-            "bound file must not warn about missing type: {:?}",
-            result.violations
-        );
-        // The skill's OWN required props are still enforced (both present here),
-        // so the file is clean.
-        assert!(
-            result
-                .violations
-                .iter()
-                .all(|v| v.severity != Severity::Error),
-            "frontmatter-less bound skill lints clean: {:?}",
-            result.violations
-        );
-    }
-
-    #[test]
-    fn bind_typed_file_still_enforces_type_specific_required() {
-        // Bind = typing drops only the `type` requirement; a bound file missing
-        // its type's OWN required property still errors.
-        let schema = schema_from_toml(
-            "\
-[schema.default]
-required = [\"type\"]
-
-[schema.types.skill]
-required = [\"name\", \"description\"]
-
-[[schema.bind]]
-glob = \"**/SKILL.md\"
-type = \"skill\"
-",
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("SKILL.md");
-        // Missing `description` (a skill-required field).
-        std::fs::write(&path, "---\nname: my-skill\n---\nBody\n").unwrap();
-
-        let result = lint_file(&path, "foo/SKILL.md", &schema, false).unwrap();
-        assert!(
-            result
-                .violations
-                .iter()
-                .any(|v| v.severity == Severity::Error
-                    && v.message
-                        .contains("missing required property \"description\"")),
-            "type-specific required still enforced: {:?}",
-            result.violations
-        );
-    }
-
-    #[test]
-    fn lint_file_no_violations_clean_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(
-            &path,
-            "---\ntitle: Hello\ntype: note\ntags:\n  - rust\n---\nBody\n",
-        )
-        .unwrap();
-
-        let schema = make_schema(&["title"], "note", &[], HashMap::new());
-        let result = lint_file(&path, "note.md", &schema, false).unwrap();
-        assert!(result.violations.is_empty());
-    }
-
-    #[test]
-    fn lint_no_schema_no_violations() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(&path, "---\ntitle: Hello\n---\nBody\n").unwrap();
-
-        let schema = SchemaConfig::default();
-        let (_, counts) = lint_extended_strict(&path, "note.md", &schema, false);
-        assert_eq!(counts.errors, 0);
-        assert_eq!(counts.warnings, 0);
-    }
-
-    // --- UX-3: comma-joined tag detection and fix ---
-
-    #[test]
-    fn lint_warns_on_comma_joined_tag() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(
-            &path,
-            "---\ntitle: Hello\ntags:\n  - cli,ux\n  - rust\n---\nBody\n",
-        )
-        .unwrap();
-
-        let schema = SchemaConfig::default();
-        let result = lint_file(&path, "note.md", &schema, false).unwrap();
-        let comma_warn = result
-            .violations
-            .iter()
-            .find(|v| v.severity == Severity::Warn && v.message.contains("cli,ux"));
-        assert!(
-            comma_warn.is_some(),
-            "expected a warning about comma-joined tag, got: {:#?}",
-            result.violations
-        );
-        assert!(
-            comma_warn.unwrap().message.contains("comma-joined"),
-            "message should mention comma-joined"
-        );
-    }
-
-    // --- Strict mode unit tests ---
-
-    fn make_schema_with_declared_prop() -> SchemaConfig {
-        // Schema with a declared `note` type that has `title` as required and
-        // `date` as a declared property, so any other property is "undeclared".
-        make_schema(&["title"], "note", &[], HashMap::new())
-    }
-
-    /// Helper: run lint in extended mode on a single file.
-    fn lint_extended_strict(
-        path: &std::path::Path,
-        rel: &str,
-        schema: &SchemaConfig,
-        strict: bool,
-    ) -> (crate::output::CommandOutcome, LintCounts) {
-        let engine = hyalo_mdlint::HyaloLintEngine::create().unwrap();
-        let md_config = hyalo_mdlint::LintConfig::default();
-        let files = vec![(path.to_path_buf(), rel.to_owned())];
-        let mut snapshot: Option<hyalo_core::index::SnapshotIndex> = None;
-        let vault_dir = path.parent().unwrap();
-        let mut opts = ExtLintOptions {
-            fix: FixMode::Off,
-            detailed: false,
-            rule_filter: None,
-            rule_prefix: None,
-            max_per_rule: 100,
-            max_files: 100,
-            fix_rules: &[],
-            snapshot_index: &mut snapshot,
-            index_path: None,
-            vault_dir,
-            strict,
-            okf_profile: false,
-            madr_profile: false,
-            skills_profile: false,
-            changelog_profile: false,
-            case_insensitive: false,
-            link_lint_ctx: None,
-            files_ignored: 0,
-        };
-        lint_files_extended(&files, schema, &engine, &md_config, &mut opts).unwrap()
-    }
-
-    /// In strict mode the "no 'type' property" warning becomes an error.
-    #[test]
-    fn strict_mode_promotes_no_type_to_error() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("no_type.md");
-        std::fs::write(&path, "---\ntitle: Hello\n---\nBody\n").unwrap();
-
-        let schema = make_schema_with_declared_prop();
-        let (_, counts) = lint_extended_strict(&path, "no_type.md", &schema, true);
-        assert!(counts.errors > 0, "strict mode: no-type should be an error");
-    }
-
-    /// Without strict mode the "no 'type' property" is still a warning.
-    #[test]
-    fn non_strict_no_type_stays_warn() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("no_type.md");
-        std::fs::write(&path, "---\ntitle: Hello\n---\nBody\n").unwrap();
-
-        let schema = make_schema_with_declared_prop();
-        let (_, counts) = lint_extended_strict(&path, "no_type.md", &schema, false);
-        assert_eq!(
-            counts.errors, 0,
-            "non-strict: no-type should remain a warning"
-        );
-        assert!(counts.warnings > 0, "non-strict: warnings expected");
-    }
-
-    /// In strict mode, undeclared properties become errors.
-    #[test]
-    fn strict_mode_promotes_undeclared_property_to_error() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("undeclared.md");
-        // `type: note` present (avoids no-type warning), but `unknown_prop` is not
-        // declared in the note schema's `properties` map.
-        std::fs::write(
-            &path,
-            "---\ntitle: Hello\ntype: note\nunknown_prop: oops\n---\nBody\n",
-        )
-        .unwrap();
-
-        let schema = {
-            // Build a schema where `note` has declared `properties` so
-            // the undeclared-property path fires.
-            use hyalo_core::schema::{PropertyConstraint, SchemaConfig, TypeSchema};
-            let mut schema = SchemaConfig::default();
-            let mut ts = TypeSchema::default();
-            ts.required.push("title".to_owned());
-            ts.properties.insert(
-                "title".to_owned(),
-                PropertyConstraint::String {
-                    pattern: None,
-                    min_length: None,
-                    max_length: None,
-                },
-            );
-            schema.types.insert("note".to_owned(), ts);
-            schema
-        };
-
-        let (_, counts) = lint_extended_strict(&path, "undeclared.md", &schema, true);
-        assert!(
-            counts.errors > 0,
-            "strict: undeclared prop should be an error"
-        );
-    }
-
-    /// A required property whose value is an empty `[]` is an error: an empty
-    /// list is semantically equivalent to absent for a required field. The
-    /// rule is value-shape driven and fires whether or not the property has a
-    /// List constraint declared.
-    #[test]
-    fn required_property_empty_array_is_error() {
-        use hyalo_core::schema::{PropertyConstraint, SchemaConfig, TypeSchema};
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("empty_tags.md");
-        std::fs::write(
-            &path,
-            "---\ntitle: Hello\ntype: note\ntags: []\n---\nBody\n",
-        )
-        .unwrap();
-
-        let schema = {
-            let mut schema = SchemaConfig::default();
-            let mut ts = TypeSchema::default();
-            ts.required.push("title".to_owned());
-            ts.required.push("tags".to_owned());
-            ts.properties
-                .insert("tags".to_owned(), PropertyConstraint::List);
-            schema.types.insert("note".to_owned(), ts);
-            schema
-        };
-
-        let (outcome, counts) = lint_extended_strict(&path, "empty_tags.md", &schema, false);
-        assert!(counts.errors > 0, "empty required list should error");
-        let body = match outcome {
-            crate::output::CommandOutcome::Success { output, .. } => output,
-            other => panic!("expected Success outcome, got: {other:?}"),
-        };
-        assert!(
-            body.contains("must not be empty") && body.contains("tags"),
-            "expected empty-required error mentioning tags in output, got: {body}"
-        );
-    }
-
-    /// A required property explicitly set to YAML null (`tags: ~`) is also
-    /// treated as empty — null carries no information, same as an absent key.
-    /// Without this, a typo or stripped value silently passes the required gate.
-    #[test]
-    fn required_property_null_value_is_error() {
-        use hyalo_core::schema::{SchemaConfig, TypeSchema};
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("null_tags.md");
-        std::fs::write(&path, "---\ntitle: Hello\ntype: note\ntags: ~\n---\nBody\n").unwrap();
-
-        let schema = {
-            let mut schema = SchemaConfig::default();
-            let mut ts = TypeSchema::default();
-            ts.required.push("tags".to_owned());
-            schema.types.insert("note".to_owned(), ts);
-            schema
-        };
-
-        let (outcome, counts) = lint_extended_strict(&path, "null_tags.md", &schema, false);
-        assert!(counts.errors > 0, "null required property should error");
-        let body = match outcome {
-            crate::output::CommandOutcome::Success { output, .. } => output,
-            other => panic!("expected Success outcome, got: {other:?}"),
-        };
-        assert!(
-            body.contains("must not be empty") && body.contains("tags"),
-            "expected empty-required error mentioning tags in output, got: {body}"
-        );
-    }
-
-    /// A required atomic-typed property satisfied by any value (including a
-    /// zero-ish one like `0` or `""`) is *not* an error from the
-    /// non-empty-list check — only sequence-typed required properties get the
-    /// extra emptiness gate.
-    #[test]
-    fn required_non_list_property_is_unaffected_by_empty_check() {
-        use hyalo_core::schema::{PropertyConstraint, SchemaConfig, TypeSchema};
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("empty_title.md");
-        std::fs::write(&path, "---\ntitle: \"\"\ntype: note\n---\nBody\n").unwrap();
-
-        let schema = {
-            let mut schema = SchemaConfig::default();
-            let mut ts = TypeSchema::default();
-            ts.required.push("title".to_owned());
-            ts.properties.insert(
-                "title".to_owned(),
-                PropertyConstraint::String {
-                    pattern: None,
-                    min_length: None,
-                    max_length: None,
-                },
-            );
-            schema.types.insert("note".to_owned(), ts);
-            schema
-        };
-
-        let (_, counts) = lint_extended_strict(&path, "empty_title.md", &schema, false);
-        assert_eq!(
-            counts.errors, 0,
-            "empty required string is not flagged here"
-        );
-    }
-
-    /// A file with `type` but no `tags` produces zero violations against a
-    /// schema that doesn't require `tags`. The previously-hardcoded "no tags
-    /// defined" warning was removed in iter-156 — opt in via `required` if you
-    /// want enforcement.
-    #[test]
-    fn missing_tags_is_not_a_violation_by_default() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("no_tags.md");
-        std::fs::write(&path, "---\ntitle: Hello\ntype: note\n---\nBody\n").unwrap();
-
-        let schema = make_schema_with_declared_prop();
-        let (_, counts) = lint_extended_strict(&path, "no_tags.md", &schema, false);
-        assert_eq!(counts.errors, 0, "missing tags should not be an error");
-        assert_eq!(counts.warnings, 0, "missing tags should not warn");
-    }
-
-    #[test]
-    fn lint_fix_splits_comma_joined_tags() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("note.md");
-        std::fs::write(
-            &path,
-            "---\ntitle: Hello\ntags:\n  - cli,ux\n  - rust\n---\nBody\n",
-        )
-        .unwrap();
-
-        let schema = SchemaConfig::default();
-        let engine = hyalo_mdlint::HyaloLintEngine::create().unwrap();
-        let md_config = hyalo_mdlint::LintConfig::default();
-        let files = vec![(path.clone(), "note.md".to_owned())];
-        let mut snapshot: Option<hyalo_core::index::SnapshotIndex> = None;
-        let mut opts = ExtLintOptions {
-            fix: FixMode::Apply,
-            detailed: false,
-            rule_filter: None,
-            rule_prefix: None,
-            max_per_rule: 100,
-            max_files: 100,
-            fix_rules: &[],
-            snapshot_index: &mut snapshot,
-            index_path: None,
-            vault_dir: dir.path(),
-            strict: false,
-            okf_profile: false,
-            madr_profile: false,
-            skills_profile: false,
-            changelog_profile: false,
-            case_insensitive: false,
-            link_lint_ctx: None,
-            files_ignored: 0,
-        };
-        let (_, counts) =
-            lint_files_extended(&files, &schema, &engine, &md_config, &mut opts).unwrap();
-
-        // After fix, the comma-joined tag warning should be gone.
-        assert_eq!(counts.warnings, 0, "comma-tag warning should be fixed");
-
-        let content = std::fs::read_to_string(&path).unwrap();
-        // Both parts of the split tag should be separate items.
-        assert!(content.contains("- cli"), "expected 'cli' as separate tag");
-        assert!(content.contains("- ux"), "expected 'ux' as separate tag");
-        // The original comma-joined form must be gone.
-        assert!(
-            !content.contains("cli,ux"),
-            "comma-joined tag should be removed"
-        );
-    }
-
-    // ---------------------------------------------------------------------------
-    // item_pattern tests
-    // ---------------------------------------------------------------------------
-
-    #[test]
-    fn item_pattern_validates_list_items() {
-        // First item matches, second does not.
-        let constraint = PropertyConstraint::StringList {
-            item_pattern: Some(r"^[a-z]+$".to_owned()),
-        };
-        let v = vc(
-            "tags",
-            &Value::Array(vec![
-                Value::String("rust".into()),
-                Value::String("Rust123".into()), // uppercase — should fail
-            ]),
-            &constraint,
-        );
-        let viol = v.expect("expected a violation");
-        assert_eq!(viol.severity, Severity::Error);
-        assert!(
-            viol.message.contains("item 1"),
-            "expected item index in message, got: {}",
-            viol.message
-        );
-        assert!(
-            viol.message.contains(r"^[a-z]+$"),
-            "expected pattern in message, got: {}",
-            viol.message
-        );
-    }
-
-    #[test]
-    fn item_pattern_vacuous_on_empty_list() {
-        let constraint = PropertyConstraint::StringList {
-            item_pattern: Some(r"^[a-z]+$".to_owned()),
-        };
-        let v = vc("tags", &Value::Array(vec![]), &constraint);
-        assert!(v.is_none(), "empty list should produce no violations");
-    }
-
-    #[test]
-    fn item_pattern_non_string_item_errors() {
-        let constraint = PropertyConstraint::StringList { item_pattern: None };
-        let v = vc(
-            "tags",
-            &Value::Array(vec![Value::Number(42.into())]),
-            &constraint,
-        );
-        let viol = v.expect("expected a violation");
-        assert_eq!(viol.severity, Severity::Error);
-        assert!(
-            viol.message.contains("item 0"),
-            "expected item index in message, got: {}",
-            viol.message
-        );
-        assert!(
-            viol.message.contains("expected string"),
-            "expected type error message, got: {}",
-            viol.message
-        );
-    }
-
-    #[test]
-    fn item_pattern_reports_all_violations() {
-        // Three items: first valid, second and third fail the pattern.
-        let constraint = PropertyConstraint::StringList {
-            item_pattern: Some(r"^[a-z][a-z0-9-]*$".to_owned()),
-        };
-        let violations = vc_all(
-            "tags",
-            &Value::Array(vec![
-                Value::String("good-tag".into()),
-                Value::String("Bad".into()),  // uppercase start — fails
-                Value::String("1bad".into()), // digit start — fails
-                Value::String("also-good".into()),
-                Value::String("Bar".into()), // uppercase start — fails
-            ]),
-            &constraint,
-        );
-        assert_eq!(
-            violations.len(),
-            3,
-            "expected 3 violations, got: {:?}",
-            violations.iter().map(|v| &v.message).collect::<Vec<_>>()
-        );
-        assert!(
-            violations[0].message.contains("item 1"),
-            "first violation should reference item 1"
-        );
-        assert!(
-            violations[1].message.contains("item 2"),
-            "second violation should reference item 2"
-        );
-        assert!(
-            violations[2].message.contains("item 4"),
-            "third violation should reference item 4"
-        );
-        for v in &violations {
-            assert_eq!(v.severity, Severity::Error);
-        }
-    }
-
-    #[test]
-    fn item_pattern_multiple_non_string_items_all_reported() {
-        // Without item_pattern, multiple non-string items should all be reported.
-        let constraint = PropertyConstraint::StringList { item_pattern: None };
-        let violations = vc_all(
-            "tags",
-            &Value::Array(vec![
-                Value::String("ok".into()),
-                Value::Number(1.into()),
-                Value::Bool(true),
-            ]),
-            &constraint,
-        );
-        assert_eq!(
-            violations.len(),
-            2,
-            "expected 2 violations for the two non-string items, got: {:?}",
-            violations.iter().map(|v| &v.message).collect::<Vec<_>>()
-        );
-        assert!(violations[0].message.contains("item 1"));
-        assert!(violations[1].message.contains("item 2"));
-    }
-
-    // ---------------------------------------------------------------------------
-    // required_sections tests
-    // ---------------------------------------------------------------------------
-
-    fn make_schema_with_sections(sections: Vec<String>) -> SchemaConfig {
-        let type_schema = TypeSchema {
-            required_sections: sections,
-            ..Default::default()
-        };
-        let mut types = HashMap::new();
-        types.insert("doc".to_owned(), type_schema);
-        SchemaConfig {
-            default: TypeSchema::default(),
-            types,
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn required_sections_pass_when_all_present() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("doc.md");
-        std::fs::write(
-            &path,
-            "---\ntype: doc\ntitle: Hello\n---\n# Goal\n\nSome text.\n\n## Tasks\n\nDo stuff.\n",
-        )
-        .unwrap();
-
-        let schema = make_schema_with_sections(vec!["# Goal".to_owned(), "## Tasks".to_owned()]);
-        let result = lint_file(&path, "doc.md", &schema, false).unwrap();
-        let section_viols: Vec<_> = result
-            .violations
-            .iter()
-            .filter(|v| v.message.contains("missing required section"))
-            .collect();
-        assert!(
-            section_viols.is_empty(),
-            "expected no section violations, got: {section_viols:?}"
-        );
-    }
-
-    #[test]
-    fn required_sections_violation_missing() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("doc.md");
-        std::fs::write(
-            &path,
-            "---\ntype: doc\ntitle: Hello\n---\n# Goal\n\nSome text.\n",
-        )
-        .unwrap();
-
-        let schema = make_schema_with_sections(vec!["# Goal".to_owned(), "## Tasks".to_owned()]);
-        let result = lint_file(&path, "doc.md", &schema, false).unwrap();
-        let section_viols: Vec<_> = result
-            .violations
-            .iter()
-            .filter(|v| v.message.contains("missing required section"))
-            .collect();
-        assert_eq!(
-            section_viols.len(),
-            1,
-            "expected exactly one missing-section violation"
-        );
-        assert!(
-            section_viols[0].message.contains("## Tasks"),
-            "expected '## Tasks' in message, got: {}",
-            section_viols[0].message
-        );
-    }
-
-    #[test]
-    fn required_sections_order_significant() {
-        // Body has both headings but in reverse order.
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("doc.md");
-        std::fs::write(
-            &path,
-            "---\ntype: doc\ntitle: Hello\n---\n## Tasks\n\nDo stuff.\n\n# Goal\n\nSome text.\n",
-        )
-        .unwrap();
-
-        // Required: Goal then Tasks (but in body: Tasks then Goal).
-        let schema = make_schema_with_sections(vec!["# Goal".to_owned(), "## Tasks".to_owned()]);
-        let result = lint_file(&path, "doc.md", &schema, false).unwrap();
-        let section_viols: Vec<_> = result
-            .violations
-            .iter()
-            .filter(|v| v.message.contains("missing required section"))
-            .collect();
-        // "# Goal" is never matched because its cursor position (after ## Tasks) is after where it appears.
-        assert!(
-            !section_viols.is_empty(),
-            "expected section violation when order is wrong"
-        );
-    }
-
-    #[test]
-    fn required_sections_extras_allowed() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("doc.md");
-        std::fs::write(
-            &path,
-            "---\ntype: doc\ntitle: Hello\n---\n# Goal\n\n## Extra One\n\nText.\n\n## Tasks\n\nDo stuff.\n\n## Extra Two\n\nMore.\n",
-        )
-        .unwrap();
-
-        let schema = make_schema_with_sections(vec!["# Goal".to_owned(), "## Tasks".to_owned()]);
-        let result = lint_file(&path, "doc.md", &schema, false).unwrap();
-        let section_viols: Vec<_> = result
-            .violations
-            .iter()
-            .filter(|v| v.message.contains("missing required section"))
-            .collect();
-        assert!(
-            section_viols.is_empty(),
-            "extra headings should not cause violations, got: {section_viols:?}"
-        );
-    }
-
-    // --- apply_body_fixes ---
-
-    fn mk_diag(
-        rule_id: &str,
-        severity: hyalo_mdlint::DiagSeverity,
-        start: usize,
-        end: usize,
-        replacement: &str,
-    ) -> hyalo_mdlint::Diagnostic {
-        hyalo_mdlint::Diagnostic {
-            rule_id: rule_id.to_owned(),
-            rule_name: rule_id.to_owned(),
-            message: String::new(),
-            line: 1,
-            column: 1,
-            severity,
-            fix: Some(hyalo_mdlint::DiagFix {
-                description: String::new(),
-                start,
-                end,
-                replacement: replacement.to_owned(),
-            }),
-        }
-    }
-
-    #[test]
-    fn apply_body_fixes_error_wins_overlap_regardless_of_offset() {
-        // Two fixes over the same range: a Warn fix that would sort first by
-        // descending-offset (its start is >= the Error fix's start) must not
-        // beat the overlapping Error fix.
-        let warn = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 0, 10, "warn-fix");
-        let error = mk_diag(
-            "HYALO001",
-            hyalo_mdlint::DiagSeverity::Error,
-            0,
-            10,
-            "error-fix",
-        );
-        let (result, outcomes) = apply_body_fixes("0123456789", &[&warn, &error]);
-        assert_eq!(result, "error-fix");
-        assert!(matches!(outcomes[1], FixOutcome::Applied));
-        assert!(matches!(outcomes[0], FixOutcome::Conflict { .. }));
-    }
-
-    #[test]
-    fn apply_body_fixes_no_op_replacement_is_not_applied() {
-        let noop = mk_diag("MD047", hyalo_mdlint::DiagSeverity::Warn, 4, 5, "\n");
-        let (result, outcomes) = apply_body_fixes("body\n", &[&noop]);
-        assert_eq!(result, "body\n", "content must not change");
-        assert!(
-            matches!(outcomes[0], FixOutcome::NoFix),
-            "byte-for-byte no-op must not be reported as Applied: {:?}",
-            outcomes[0]
-        );
-    }
-
-    #[test]
-    fn apply_body_fixes_rejects_inverted_range() {
-        // start > end: `body[start..end]` would panic. The fix must be
-        // reported as a conflict and the body left untouched (iter-191).
-        let inverted = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 4, 1, "boom");
-        let (result, outcomes) = apply_body_fixes("0123456789", &[&inverted]);
-        assert_eq!(result, "0123456789", "content must not change");
-        assert!(
-            matches!(outcomes[0], FixOutcome::Conflict { .. }),
-            "inverted range must be a Conflict, got: {:?}",
-            outcomes[0]
-        );
-    }
-
-    #[test]
-    fn apply_body_fixes_rejects_non_char_boundary() {
-        // "é" is two bytes; end == 1 lands mid-character, which `str`
-        // indexing rejects with a panic. In-bounds, so the old
-        // `end > body.len()` check missed it entirely (iter-191).
-        let body = "é abc";
-        let mid_char = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 0, 1, "x");
-        let (result, outcomes) = apply_body_fixes(body, &[&mid_char]);
-        assert_eq!(result, body, "content must not change");
-        assert!(
-            matches!(outcomes[0], FixOutcome::Conflict { .. }),
-            "mid-char-boundary range must be a Conflict, got: {:?}",
-            outcomes[0]
-        );
-    }
-
-    #[test]
-    fn apply_body_fixes_non_overlapping_fixes_both_apply() {
-        let a = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 0, 1, "A");
-        let b = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 2, 3, "B");
-        let (result, outcomes) = apply_body_fixes("xyz", &[&a, &b]);
-        assert_eq!(result, "AyB");
-        assert!(matches!(outcomes[0], FixOutcome::Applied));
-        assert!(matches!(outcomes[1], FixOutcome::Applied));
-    }
-
-    #[test]
-    fn apply_body_fixes_adjacent_touching_ranges_both_apply() {
-        // end == start is NOT an overlap (strict inequalities in the
-        // conflict check): [0,2) and [2,4) must both apply.
-        let a = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 0, 2, "AA");
-        let b = mk_diag("MD009", hyalo_mdlint::DiagSeverity::Warn, 2, 4, "BB");
-        let (result, outcomes) = apply_body_fixes("wxyz", &[&a, &b]);
-        assert_eq!(result, "AABB");
-        assert!(matches!(outcomes[0], FixOutcome::Applied));
-        assert!(matches!(outcomes[1], FixOutcome::Applied));
-    }
-
-    // --- find_body_start line-ending handling ---
-
-    #[test]
-    fn find_body_start_skips_crlf_after_closing_delimiter() {
-        let content = "---\r\ntitle: T\r\n---\r\nbody line\r\n";
-        let start = find_body_start(content);
-        assert_eq!(
-            &content[start..],
-            "body line\r\n",
-            "body must not start with a stray CR on CRLF files"
-        );
-    }
-
-    #[test]
-    fn find_body_start_bom_prefixed_frontmatter_is_split() {
-        let content = "\u{feff}---\ntitle: T\n---\nbody line\n";
-        let start = find_body_start(content);
-        assert_eq!(&content[start..], "body line\n");
-    }
-
-    // --- group_severity ---
-
-    fn iv(severity: &str) -> InternalViolation {
-        InternalViolation {
-            line: 1,
-            column: 1,
-            message: String::new(),
-            severity: severity.to_owned(),
-            fix: None,
-            fixed: false,
-            autofixable: None,
-        }
-    }
-
-    #[test]
-    fn group_severity_is_error_when_any_violation_is_error() {
-        let violations = vec![iv("warn"), iv("error"), iv("warn")];
-        assert_eq!(group_severity(&violations), "error");
-    }
-
-    #[test]
-    fn group_severity_is_warn_when_all_warn() {
-        let violations = vec![iv("warn"), iv("warn")];
-        assert_eq!(group_severity(&violations), "warn");
-    }
-
-    #[test]
-    fn group_severity_empty_defaults_to_warn() {
-        assert_eq!(group_severity(&[]), "warn");
-    }
-}
-
-/// ARCH-1 (iter-225): the `hyalo lint` dispatch arm, extracted verbatim from
-/// `dispatch.rs` so its warning policy (`[lint] ignore` notices), profile
-/// selection and exit-code logic are unit-testable in-process.
-///
-/// `profile`, `files_from` and `index_flags` were consumed earlier in
-/// `run.rs`/dispatch (profile overlay into `ctx.lint_profiles`, snapshot
-/// loading), so they never reach here.
-use crate::dispatch::{
-    CommandContext, adapt_view_result_to_ext, inject_ext_file_result, maybe_case_index,
-};
-use hyalo_core::mode_enabled;
-
-#[allow(clippy::too_many_arguments)]
-#[allow(clippy::fn_params_excessive_bools)] // moved verbatim from the dispatch arm
-#[allow(clippy::needless_pass_by_value)] // args moved verbatim from the clap variant
-pub(crate) fn run(
-    ctx: &mut CommandContext<'_>,
-    file_positional: Vec<String>,
-    file: Vec<String>,
-    glob: Vec<String>,
-    lint_type: Option<String>,
-    fix: bool,
-    dry_run: bool,
-    cli_limit: Option<usize>,
-    detailed: bool,
-    rule: Option<String>,
-    rule_prefix: Option<String>,
-    max_per_rule: Option<usize>,
-    fix_rule: Vec<String>,
-    lint_strict_flag: bool,
-) -> Result<CommandOutcome> {
-    let dir = ctx.dir;
-    let site_prefix = ctx.site_prefix;
-    let snapshot_index = &mut *ctx.snapshot_index;
-    let index_path = ctx.index_path;
-    let profile_active = |name: &str| ctx.lint_profiles.iter().any(|p| p == name);
-    let okf_profile_active = profile_active("okf");
-    let madr_profile_active = profile_active("madr");
-    let skills_profile_active = profile_active("skills");
-    let changelog_profile_active = profile_active("changelog");
-
-    {
-        // --strict flag wins over config value; config value is the fallback.
-        let effective_strict = lint_strict_flag || ctx.lint_strict;
-        // Resolve --type to a glob pattern from its filename_template.
-        let type_glob: Option<String> = if let Some(type_name) = lint_type {
-            use hyalo_core::filename_template::FilenameTemplate;
-            match ctx.schema.types.get(&type_name) {
-                Some(ts) => match &ts.filename_template {
-                    Some(template_str) => match FilenameTemplate::parse(template_str) {
-                        Ok(tpl) => Some(tpl.to_glob()),
-                        Err(e) => {
-                            return Ok(crate::output::CommandOutcome::UserError(
-                                crate::output::format_error(
-                                    ctx.user_format,
-                                    &format!(
-                                        "invalid filename_template for type '{type_name}': {e}"
-                                    ),
-                                    None,
-                                    None,
-                                    None,
-                                ),
-                            ));
-                        }
-                    },
-                    None => {
-                        return Ok(crate::output::CommandOutcome::UserError(
-                            crate::output::format_error(
-                                ctx.user_format,
-                                &format!("type '{type_name}' has no filename_template defined"),
-                                None,
-                                Some(
-                                    "set one with: hyalo types set <name> --filename-template <pattern>",
-                                ),
-                                None,
-                            ),
-                        ));
-                    }
-                },
-                None => {
-                    return Ok(crate::output::CommandOutcome::UserError(
-                        crate::output::format_error(
-                            ctx.user_format,
-                            &format!("unknown type '{type_name}'"),
-                            None,
-                            Some("run `hyalo types list` to see available types"),
-                            None,
-                        ),
-                    ));
-                }
-            }
-        } else {
-            None
-        };
-
-        // Build the file list. Positional args are treated as --file
-        // targets (repeatable), preserving their command-line order ahead
-        // of any --file values.
-        let mut files_arg: Vec<String> = file_positional;
-        files_arg.extend(file);
-        // --type expands to a glob that overrides file/glob args.
-        let effective_glob: Vec<String> = if let Some(g) = type_glob {
-            vec![g]
-        } else {
-            glob
-        };
-
-        let mut file_pairs = match crate::commands::collect_files(
-            dir,
-            &files_arg,
-            &effective_glob,
-            ctx.user_format,
-        )? {
-            crate::commands::FilesOrOutcome::Files(f) => f,
-            crate::commands::FilesOrOutcome::Outcome(o) => return Ok(o),
-        };
-
-        // Reach a repo-root CHANGELOG.md that lives *outside* the vault dir.
-        // When the changelog profile is active and `[changelog] path`
-        // resolves to a file the vault walk can't see (the common
-        // docs-subdir layout), add it to the lint set so
-        // `lint --profile changelog` validates the real file without
-        // `--dir .` gymnastics. Only for an unscoped run (no explicit
-        // `--file`/`--glob`), so a targeted lint is never surprised by an
-        // extra file.
-        if files_arg.is_empty()
-            && effective_glob.is_empty()
-            && ctx.lint_profiles.iter().any(|p| p == "changelog")
-            && let Ok(changelog_file) = crate::commands::changelog::resolve_changelog_file(
-                dir,
-                ctx.config_dir,
-                ctx.changelog_path,
-            )
-            && changelog_file.is_file()
-        {
-            let rel = hyalo_core::discovery::relative_path(dir, &changelog_file);
-            // Only inject when the file is outside the vault dir (a relative
-            // path that climbs out, or an absolute one) — an in-vault
-            // CHANGELOG.md was already discovered by the walk.
-            let outside = rel.starts_with("..") || std::path::Path::new(&rel).is_absolute();
-            let already = file_pairs.iter().any(|(p, _)| p == &changelog_file);
-            if outside && !already {
-                let display = changelog_file
-                    .file_name()
-                    .map_or_else(|| rel.clone(), |n| n.to_string_lossy().into_owned());
-                file_pairs.push((changelog_file, display));
-            }
-        }
-
-        let fix_mode = if fix {
-            if dry_run {
-                self::FixMode::DryRun
-            } else {
-                self::FixMode::Apply
-            }
-        } else {
-            self::FixMode::Off
-        };
-
-        // Filter out files matching `[lint] ignore` entries.
-        //
-        // Each entry is matched against the vault-relative path (with `/`
-        // separators) as a glob: `vendor/**/*.md`, `legacy/known-bad.md`,
-        // `templates/*.md`. An entry without glob meta-characters is matched
-        // literally (exact path equality on the normalized path).
-        // `--file`/positional args are "explicit": the user named them
-        // directly rather than sweeping a glob. When an explicit file is
-        // excluded by `[lint] ignore` it must produce a visible notice, not
-        // a silent `0 files checked` (df-scale silent-drop family).
-        let explicit_named = !files_arg.is_empty();
-        // UX-1 (dogfood pre3): how many files `[lint] ignore` dropped from
-        // this run, regardless of scope. Zero when there is no `[lint]
-        // ignore` at all (the branch below never runs).
-        let mut lint_ignored_count: usize = 0;
-        let filtered_pairs: Vec<_> = if ctx.lint_ignore.is_empty() {
-            file_pairs
-        } else {
-            use globset::{GlobBuilder, GlobSetBuilder};
-            let mut builder = GlobSetBuilder::new();
-            let mut build_failed = false;
-            for pat in ctx.lint_ignore {
-                match GlobBuilder::new(pat)
-                    .literal_separator(true)
-                    .backslash_escape(true)
-                    .build()
-                {
-                    Ok(g) => {
-                        builder.add(g);
-                    }
-                    Err(e) => {
-                        crate::warn::warn(format!("invalid [lint] ignore pattern {pat:?}: {e}"));
-                        build_failed = true;
-                    }
-                }
-            }
-            let set = if build_failed {
-                None
-            } else {
-                builder.build().ok()
-            };
-            // UX-1 (dogfood pre3): a `--glob` sweep is just as "explicit"
-            // as `--file` when it comes to the *all-matches-ignored* case
-            // — the user asked for a specific set of files by name or by
-            // pattern either way, and getting back an empty, vacuously
-            // green lint run with no explanation is the same silent-drop
-            // trap regardless of which form they used.
-            let explicit_glob = !effective_glob.is_empty();
-            match set {
-                Some(set) => {
-                    let mut ignored_named: Vec<String> = Vec::new();
-                    let before = file_pairs.len();
-                    let kept: Vec<_> = file_pairs
-                        .into_iter()
-                        .filter(|(_, rel)| {
-                            let norm = rel.replace('\\', "/");
-                            let matched = set.is_match(&norm);
-                            if matched && (explicit_named || explicit_glob) {
-                                ignored_named.push(norm);
-                            }
-                            !matched
-                        })
-                        .collect();
-                    // UX-1: unconditional count, regardless of how the
-                    // file set was scoped — feeds the bare-sweep summary
-                    // line ("N files checked (M ignored by [lint]
-                    // ignore)") so a full-vault run stops hiding how much
-                    // of the vault it silently skipped.
-                    lint_ignored_count = before - kept.len();
-                    // Notice for an explicit scope (named files, or a
-                    // --glob whose matches are *entirely* ignored)
-                    // silently excluded by `[lint] ignore` — otherwise the
-                    // run reports `0 files checked, no issues` with no
-                    // hint why. A --glob that only partially matches the
-                    // ignore list stays quiet here: the bare-sweep count
-                    // above already makes the exclusion visible without
-                    // the noise of naming every match.
-                    let glob_all_ignored = explicit_glob && !explicit_named && kept.is_empty();
-                    if !ignored_named.is_empty() && (explicit_named || glob_all_ignored) {
-                        let list = ignored_named.join(", ");
-                        let plural = if ignored_named.len() == 1 {
-                            "file"
-                        } else {
-                            "files"
-                        };
-                        crate::warn::warn(format!(
-                            "{} named {plural} excluded by [lint] ignore (not linted): {list}",
-                            ignored_named.len()
-                        ));
-                    }
-                    kept
-                }
-                // If building the set failed (warning already emitted above),
-                // fall back to no filtering rather than silently ignoring
-                // potentially relevant files.
-                None => file_pairs,
-            }
-        };
-
-        // Decide which lint path to use: extended (body+frontmatter) or legacy.
-        // The extended path is used whenever the new flags are active OR when the
-        // engine is available.  We always use the extended path now.
-        let md_engine = hyalo_mdlint::HyaloLintEngine::create()
-            .map_err(|e| anyhow::anyhow!("failed to create lint engine: {e}"))?;
-
-        // M-10: `--rule` names a rule id, so validate it the way
-        // `lint-rules show` does instead of silently linting with a filter
-        // that matches nothing (an unknown id used to exit 0 with "no
-        // issues found", which reads as "clean" in CI). The match is
-        // case-insensitive and canonicalizes to the catalog spelling so
-        // `--rule hyalo006` behaves exactly like `--rule HYALO006`.
-        let rule = match rule {
-            Some(raw) => match md_engine.rule_entry_ci(&raw) {
-                Some(entry) => Some(entry.id.clone()),
-                None => {
-                    return Ok(crate::output::CommandOutcome::UserError(
-                        crate::output::format_error(
-                            ctx.user_format,
-                            &format!("no such rule: {raw}"),
-                            None,
-                            Some("run `hyalo lint-rules list` to see available rules"),
-                            None,
-                        ),
-                    ));
-                }
-            },
-            None => None,
-        };
-        // iter-210 BUG-5: a prefix that selects no rule is as much a typo
-        // as an unknown `--rule` id, and it used to be *worse* than one:
-        // the empty filter fell through to "no filtering", so `--rule-prefix
-        // nope` warned and then ran every MD rule anyway at exit 0. Fail it
-        // with the same error shape as `--rule`. The match is
-        // case-insensitive, so a matching family is unaffected.
-        if let Some(prefix) = rule_prefix.as_deref()
-            && md_engine.rules_matching_prefix_ci(prefix).is_empty()
-        {
-            return Ok(crate::output::CommandOutcome::UserError(
-                crate::output::format_error(
-                    ctx.user_format,
-                    &format!("no rule matches prefix: {prefix}"),
-                    None,
-                    Some("run `hyalo lint-rules list` to see available rules"),
-                    None,
-                ),
-            ));
-        }
-
-        // `--format github` emits one annotation per violation, so every
-        // finding must be materialized: force `detailed` and lift the
-        // per-rule / per-file caps. Otherwise the summary-mode truncation
-        // would silently drop annotations from the PR check.
-        let github_output = ctx.user_format == crate::output::Format::Github;
-
-        let max_per_rule_eff = if github_output {
-            // `usize::MAX`, not `0` — the fix-mode output path uses
-            // `.take(n)`/`.min(n)` to cap violations shown per rule, so `0`
-            // would truncate every rule group to zero shown violations and
-            // silently drop annotations. The non-fix path separately bypasses
-            // this cap via `opts.detailed`, but fix-mode does not, so the
-            // sentinel must mean "unlimited" in both paths.
-            usize::MAX
-        } else {
-            max_per_rule.unwrap_or_else(|| ctx.md_lint.max_violations_per_rule())
-        };
-        // CLI --limit overrides the config max_files when provided.
-        // `--limit 0` is documented as "unlimited" (matches `--count
-        // --limit 0`): map it to `usize::MAX` so it lifts the file cap
-        // instead of truncating the list to zero (ff-rdp B5, mapl BUG-4).
-        let max_files_eff = match cli_limit {
-            Some(0) => usize::MAX,
-            Some(n) => n,
-            None if github_output => usize::MAX,
-            None => ctx.md_lint.max_files(),
-        };
-
-        // HYALO006 (broken-link): build the vault-wide resolution context
-        // ONCE per invocation, but only when the rule will actually run —
-        // i.e. it is enabled (no `[lint.rules.HYALO006] enabled = false`)
-        // AND selected by any `--rule` / `--rule-prefix` filter. Building it
-        // means seeding a case/stem index (from the snapshot when `--index`
-        // is active, else a single disk walk) — never per file.
-        let hyalo006_enabled = ctx
-            .md_lint
-            .rules
-            .get(self::RULE_ID_BROKEN_LINK)
-            .and_then(hyalo_mdlint::RuleOverride::enabled)
-            .unwrap_or(true);
-        let hyalo006_selected = match (&rule, &rule_prefix) {
-            (Some(r), _) => r == self::RULE_ID_BROKEN_LINK,
-            (None, Some(p)) => self::RULE_ID_BROKEN_LINK
-                .to_ascii_uppercase()
-                .starts_with(&p.to_ascii_uppercase()),
-            (None, None) => true,
-        };
-        let link_lint_ctx = if hyalo006_enabled && hyalo006_selected {
-            let case_index = maybe_case_index(
-                ctx.case_insensitive_mode,
-                dir,
-                true,
-                (*snapshot_index).as_ref(),
-            )
-            .unwrap_or_default();
-            hyalo_mdlint::profiles::link::LinkLintContext::new(
-                dir,
-                site_prefix.map(str::to_owned),
-                case_index,
-            )
-        } else {
-            None
-        };
-
-        let mut ext_opts = self::ExtLintOptions {
-            fix: fix_mode,
-            detailed: detailed || github_output,
-            rule_filter: rule.as_deref(),
-            rule_prefix: rule_prefix.as_deref(),
-            max_per_rule: max_per_rule_eff,
-            max_files: max_files_eff,
-            fix_rules: &fix_rule,
-            snapshot_index,
-            index_path,
-            vault_dir: dir,
-            strict: effective_strict,
-            // The active profile is resolved in run.rs (CLI `--profile`
-            // overlay OR `[lint] profile` in config); captured above.
-            okf_profile: okf_profile_active,
-            madr_profile: madr_profile_active,
-            skills_profile: skills_profile_active,
-            changelog_profile: changelog_profile_active,
-            // Resolved once per `hyalo lint` invocation (not re-probed
-            // per file) so `[schema] exempt` globs fold case the same way
-            // `hyalo okf index` treats `INDEX.md` on case-insensitive
-            // filesystems (macOS/Windows default).
-            case_insensitive: mode_enabled(ctx.case_insensitive_mode, dir),
-            link_lint_ctx,
-            files_ignored: lint_ignored_count,
-        };
-
-        let (outcome, mut counts) = self::lint_files_extended(
-            &filtered_pairs,
-            ctx.schema,
-            &md_engine,
-            ctx.md_lint,
-            &mut ext_opts,
-        )?;
-
-        // Additional config-level lint: check view definitions AND that
-        // [schema] itself parses (review round finding 2 — a malformed
-        // [schema] block used to be only a stderr warning, so `lint
-        // --strict` could exit 0 on a vault whose schema validation was
-        // silently disabled). Merged into one `.hyalo.toml` pseudo-file
-        // result so both kinds of config-level problem show up together
-        // rather than as two separate file entries.
-        let mut config_result: Option<self::FileLintResult> = self::validate_views(ctx.config_dir);
-        if let Some(schema_result) = self::validate_schema_config(ctx.config_dir, effective_strict)
-        {
-            match &mut config_result {
-                Some(existing) => existing.violations.extend(schema_result.violations),
-                None => config_result = Some(schema_result),
-            }
-        }
-        let outcome = if let Some(config_result) = config_result {
-            for v in &config_result.violations {
-                match v.severity {
-                    self::Severity::Error => counts.errors += 1,
-                    self::Severity::Warn => counts.warnings += 1,
-                }
-            }
-            counts.files_with_issues += 1;
-            // Adapt into the new shape — inject as a file with a SCHEMA group.
-            let adapted = adapt_view_result_to_ext(&config_result);
-            inject_ext_file_result(outcome, &adapted)?
-        } else {
-            outcome
-        };
-
-        // Signal exit code 1 when errors remain after fixes (set before returning).
-        if counts.errors > 0 {
-            ctx.exit_code_override = Some(1);
-        }
-
-        Ok(outcome)
-    }
-}
+mod tests;
